% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{shiny.webawesome-package}
\alias{shiny.webawesome}
\alias{shiny.webawesome-package}
\title{shiny.webawesome: Shiny Bindings for Web Awesome Components}
\description{
Provides an R and Shiny interface to the Web Awesome component library.
}
\author{
\strong{Maintainer}: M. B. Anand \email{anand@owlcyberlabs.com}

}
\keyword{internal}
