% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_animated_image.R
\name{wa_animated_image}
\alias{wa_animated_image}
\title{Create a \code{wa-animated-image} component}
\usage{
wa_animated_image(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  alt = NULL,
  dir = NULL,
  lang = NULL,
  play = NULL,
  src = NULL,
  pause_icon = NULL,
  play_icon = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{alt}{String. A description of the image used by assistive
devices.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{play}{Boolean. Default: \code{FALSE}. Plays the animation. When this
attribute is remove, the animation will pause.}

\item{src}{String. The path to the image to load.}

\item{pause_icon}{Optional pause icon to use instead of the default.
Works best with \verb{<wa-icon>}.}

\item{play_icon}{Optional play icon to use instead of the default.
Works best with \verb{<wa-icon>}.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-animated-image} component.
Generated from Web Awesome metadata.
}
