% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_animation.R
\name{wa_animation}
\alias{wa_animation}
\title{Create a \code{wa-animation} component}
\usage{
wa_animation(
  ...,
  id = NULL,
  name = NULL,
  delay = NULL,
  dir = NULL,
  direction = NULL,
  duration = NULL,
  easing = NULL,
  end_delay = NULL,
  fill = NULL,
  iteration_start = NULL,
  iterations = NULL,
  lang = NULL,
  play = NULL,
  playback_rate = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{name}{The name of the built-in animation to use. For custom
animations, use the \code{keyframes} prop. Defaults to \code{none} when omitted.}

\item{delay}{The number of milliseconds to delay the start of the
animation. Defaults to \code{0} when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{direction}{Determines the direction of playback as well as the
behavior when reaching the end of an iteration. Learn more Defaults to
\code{normal} when omitted.}

\item{duration}{The number of milliseconds each iteration of the
animation takes to complete. Defaults to \code{1000} when omitted.}

\item{easing}{The easing function to use for the animation. This can be
a Web Awesome easing function or a custom easing function such as
\code{cubic-bezier(0, 1, .76, 1.14)}. Defaults to \code{linear} when omitted.}

\item{end_delay}{The number of milliseconds to delay after the active
period of an animation sequence. Defaults to \code{0} when omitted.}

\item{fill}{Sets how the animation applies styles to its target before
and after its execution. Defaults to \code{auto} when omitted.}

\item{iteration_start}{The offset at which to start the animation,
usually between 0 (start) and 1 (end). Defaults to \code{0} when omitted.}

\item{iterations}{The number of iterations to run before the animation
completes. Defaults to \code{Infinity}, which loops. Defaults to \code{Infinity}
when omitted.}

\item{lang}{Optional Web Awesome attribute.}

\item{play}{Plays the animation. When omitted, the animation will be
paused. This attribute will be automatically removed when the animation
finishes or gets canceled. Defaults to \code{false} when omitted.}

\item{playback_rate}{Sets the animation's playback rate. The default is
\code{1}, which plays the animation at a normal speed. Setting this to \code{2},
for example, will double the animation's speed. A negative value can be
used to reverse the animation. This value can be changed without
causing the animation to restart. Defaults to \code{1} when omitted.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-animation} component.
Generated from Web Awesome metadata.
}
