% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_avatar.R
\name{wa_avatar}
\alias{wa_avatar}
\title{Create a \code{wa-avatar} component}
\usage{
wa_avatar(
  ...,
  id = NULL,
  label = NULL,
  dir = NULL,
  image = NULL,
  initials = NULL,
  lang = NULL,
  loading = NULL,
  shape = NULL,
  icon = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{label}{A label to use to describe the avatar to assistive
devices. Defaults to `` when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{image}{The image source to use for the avatar. Defaults to ``
when omitted.}

\item{initials}{Initials to use as a fallback when no image is
available (1-2 characters max recommended). Defaults to `` when
omitted.}

\item{lang}{Optional Web Awesome attribute.}

\item{loading}{Indicates how the browser should load the image. Must be
one of \code{"eager"}, \code{"lazy"}. Defaults to \code{eager} when omitted.}

\item{shape}{The shape of the avatar. Must be one of \code{"circle"},
\code{"rounded"}, \code{"square"}. Defaults to \code{circle} when omitted.}

\item{icon}{The default icon to use when no image or initials are
present. Works best with \verb{<wa-icon>}.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-avatar} component.
Generated from Web Awesome metadata.
}
