% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_badge.R
\name{wa_badge}
\alias{wa_badge}
\title{Create a \code{wa-badge} component}
\usage{
wa_badge(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  appearance = NULL,
  attention = NULL,
  dir = NULL,
  lang = NULL,
  pill = NULL,
  variant = NULL,
  end = NULL,
  start = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{appearance}{Enumerated string. Allowed values: \code{accent},
\code{filled}, \code{filled-outlined}, \code{outlined}. Default: \code{accent}. The badge's
visual appearance.}

\item{attention}{Enumerated string. Allowed values: \code{bounce}, \code{none},
\code{pulse}. Default: \code{none}. Adds an animation to draw attention to the
badge.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{pill}{Boolean. Default: \code{FALSE}. Draws a pill-style badge with
rounded edges.}

\item{variant}{Enumerated string. Allowed values: \code{brand}, \code{danger},
\code{neutral}, \code{success}, \code{warning}. Default: \code{brand}. The badge's theme
variant. Defaults to \code{brand} if not within another element with a
variant.}

\item{end}{An element, such as \verb{<wa-icon>}, placed after the label.}

\item{start}{An element, such as \verb{<wa-icon>}, placed before the label.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-badge} component.
Generated from Web Awesome metadata.
}
