% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_breadcrumb.R
\name{wa_breadcrumb}
\alias{wa_breadcrumb}
\title{Create a \code{wa-breadcrumb} component}
\usage{
wa_breadcrumb(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  label = NULL,
  dir = NULL,
  lang = NULL,
  separator = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{label}{String. Default: \code{""}. The label to use for the breadcrumb
control. This will not be shown on the screen, but it will be announced
by screen readers and other assistive devices to provide more context
for users.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{separator}{The separator to use between breadcrumb items. Works
best with \verb{<wa-icon>}.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-breadcrumb} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

