% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_breadcrumb_item.R
\name{wa_breadcrumb_item}
\alias{wa_breadcrumb_item}
\title{Create a \code{wa-breadcrumb-item} component}
\usage{
wa_breadcrumb_item(
  ...,
  id = NULL,
  dir = NULL,
  href = NULL,
  lang = NULL,
  rel = NULL,
  target = NULL,
  end = NULL,
  separator = NULL,
  start = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{dir}{Optional Web Awesome attribute.}

\item{href}{Optional URL to direct the user to when the breadcrumb item
is activated. When set, a link will be rendered internally. When unset,
a button will be rendered instead.}

\item{lang}{Optional Web Awesome attribute.}

\item{rel}{The \code{rel} attribute to use on the link. Only used when
\code{href} is set. Defaults to \verb{noreferrer noopener} when omitted.}

\item{target}{Tells the browser where to open the link. Only used when
\code{href} is set.}

\item{end}{An element, such as \verb{<wa-icon>}, placed after the label.}

\item{separator}{The separator to use for the breadcrumb item. This
will only change the separator for this item. If you want to change it
for all items in the group, set the separator on \verb{<wa-breadcrumb>}
instead.}

\item{start}{An element, such as \verb{<wa-icon>}, placed before the label.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-breadcrumb-item} component.
Generated from Web Awesome metadata.
}
