% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_button.R
\name{wa_button}
\alias{wa_button}
\title{Create a \code{wa-button} component}
\usage{
wa_button(
  ...,
  id = NULL,
  value = NULL,
  disabled = NULL,
  name = NULL,
  appearance = NULL,
  custom_error = NULL,
  dir = NULL,
  download = NULL,
  formaction = NULL,
  formenctype = NULL,
  formmethod = NULL,
  formnovalidate = NULL,
  formtarget = NULL,
  href = NULL,
  lang = NULL,
  loading = NULL,
  pill = NULL,
  rel = NULL,
  size = NULL,
  target = NULL,
  title = NULL,
  type = NULL,
  variant = NULL,
  with_caret = NULL,
  end = NULL,
  start = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{value}{The value of the button, submitted as a pair with the
button's name as part of the form data, but only when this button is
the submitter. This attribute is ignored when \code{href} is present.}

\item{disabled}{Disables the button.}

\item{name}{The name of the button, submitted as a name/value pair with
form data, but only when this button is the submitter. This attribute
is ignored when \code{href} is present.}

\item{appearance}{The button's visual appearance.}

\item{custom_error}{Optional Web Awesome attribute.}

\item{dir}{Optional Web Awesome attribute.}

\item{download}{Tells the browser to download the linked file as this
filename. Only used when \code{href} is present.}

\item{formaction}{Used to override the form owner's \code{action} attribute.}

\item{formenctype}{Used to override the form owner's \code{enctype}
attribute.}

\item{formmethod}{Used to override the form owner's \code{method} attribute.}

\item{formnovalidate}{Used to override the form owner's \code{novalidate}
attribute.}

\item{formtarget}{Used to override the form owner's \code{target} attribute.}

\item{href}{When set, the underlying button will be rendered as an
\verb{<a>} with this \code{href} instead of a \verb{<button>}.}

\item{lang}{Optional Web Awesome attribute.}

\item{loading}{Draws the button in a loading state.}

\item{pill}{Draws a pill-style button with rounded edges.}

\item{rel}{When using \code{href}, this attribute will map to the underlying
link's \code{rel} attribute.}

\item{size}{The button's size.}

\item{target}{Tells the browser where to open the link. Only used when
\code{href} is present.}

\item{title}{Optional Web Awesome attribute.}

\item{type}{The type of button. Note that the default value is \code{button}
instead of \code{submit}, which is opposite of how native \verb{<button>}
elements behave. When the type is \code{submit}, the button will submit the
surrounding form.}

\item{variant}{The button's theme variant. Defaults to \code{neutral} if not
within another element with a variant.}

\item{with_caret}{Draws the button with a caret. Used to indicate that
the button triggers a dropdown menu or similar behavior.}

\item{end}{An element, such as \verb{<wa-icon>}, placed after the label.}

\item{start}{An element, such as \verb{<wa-icon>}, placed before the label.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-button} component.
Generated from Web Awesome metadata.
}
