% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_button_group.R
\name{wa_button_group}
\alias{wa_button_group}
\title{Create a \code{wa-button-group} component}
\usage{
wa_button_group(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  label = NULL,
  dir = NULL,
  lang = NULL,
  orientation = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{label}{String. Default: \code{""}. A label to use for the button
group. This won't be displayed on the screen, but it will be announced
by assistive devices when interacting with the control and is strongly
recommended.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{orientation}{Enumerated string. Allowed values: \code{horizontal},
\code{vertical}. Default: \code{horizontal}. The button group's orientation.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-button-group} component.
Generated from Web Awesome metadata.
}
