% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_callout.R
\name{wa_callout}
\alias{wa_callout}
\title{Create a \code{wa-callout} component}
\usage{
wa_callout(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  appearance = NULL,
  dir = NULL,
  lang = NULL,
  size = NULL,
  variant = NULL,
  icon = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{appearance}{Enumerated string. Allowed values: \code{accent},
\code{filled}, \code{filled-outlined}, \code{outlined}, \code{plain}. The callout's visual
appearance.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{size}{Enumerated string. Allowed values: \code{large}, \code{medium},
\code{small}. Default: \code{medium}. The callout's size.}

\item{variant}{Enumerated string. Allowed values: \code{brand}, \code{danger},
\code{neutral}, \code{success}, \code{warning}. Default: \code{brand}. The callout's theme
variant. Defaults to \code{brand} if not within another element with a
variant.}

\item{icon}{An icon to show in the callout. Works best with
\verb{<wa-icon>}.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-callout} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

