% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_checkbox.R
\name{wa_checkbox}
\alias{wa_checkbox}
\title{Create a \code{wa-checkbox} component}
\usage{
wa_checkbox(
  ...,
  id = NULL,
  value = NULL,
  checked = NULL,
  disabled = NULL,
  hint = NULL,
  name = NULL,
  custom_error = NULL,
  dir = NULL,
  indeterminate = NULL,
  lang = NULL,
  required = NULL,
  size = NULL,
  title = NULL,
  hint_slot = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute.}

\item{value}{The value of the checkbox, submitted as a name/value pair with form data.}

\item{checked}{The default value of the form control. Primarily used for resetting the form control.}

\item{disabled}{Disables the checkbox.}

\item{hint}{The checkbox's hint. If you need to display HTML, use the \code{hint} slot instead.}

\item{name}{The name of the checkbox, submitted as a name/value pair with form data.}

\item{custom_error}{Optional Web Awesome attribute.}

\item{dir}{Optional Web Awesome attribute.}

\item{indeterminate}{Draws the checkbox in an indeterminate state. This is usually applied to checkboxes that represents a "select all/none" behavior when associated checkboxes have a mix of checked and unchecked states.}

\item{lang}{Optional Web Awesome attribute.}

\item{required}{Makes the checkbox a required field.}

\item{size}{The checkbox's size.}

\item{title}{Optional Web Awesome attribute.}

\item{hint_slot}{Text that describes how to use the checkbox. Alternatively, you can use the \code{hint} attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-checkbox} component.
Generated from Web Awesome metadata.
}
