% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_color_picker.R
\name{wa_color_picker}
\alias{wa_color_picker}
\title{Create a \code{wa-color-picker} component}
\usage{
wa_color_picker(
  input_id,
  ...,
  value = NULL,
  disabled = NULL,
  label = NULL,
  hint = NULL,
  name = NULL,
  custom_error = NULL,
  dir = NULL,
  format = NULL,
  lang = NULL,
  opacity = NULL,
  open = NULL,
  required = NULL,
  size = NULL,
  swatches = NULL,
  uppercase = NULL,
  with_hint = NULL,
  with_label = NULL,
  without_format_toggle = NULL,
  hint_slot = NULL,
  label_slot = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{value}{The default value of the form control. Primarily used for
resetting the form control.}

\item{disabled}{Disables the color picker.}

\item{label}{The color picker's label. This will not be displayed, but
it will be announced by assistive devices. If you need to display HTML,
you can use the \code{label} slot` instead.}

\item{hint}{The color picker's hint. If you need to display HTML, use
the \code{hint} slot instead.}

\item{name}{The name of the form control, submitted as a name/value
pair with form data.}

\item{custom_error}{Optional Web Awesome attribute.}

\item{dir}{Optional Web Awesome attribute.}

\item{format}{The format to use. If opacity is enabled, these will
translate to HEXA, RGBA, HSLA, and HSVA respectively. The color picker
will accept user input in any format (including CSS color names) and
convert it to the desired format.}

\item{lang}{Optional Web Awesome attribute.}

\item{opacity}{Shows the opacity slider. Enabling this will cause the
formatted value to be HEXA, RGBA, or HSLA.}

\item{open}{Indicates whether or not the popup is open. You can toggle
this attribute to show and hide the popup, or you can use the \code{show()}
and \code{hide()} methods and this attribute will reflect the popup's open
state.}

\item{required}{Makes the color picker a required field.}

\item{size}{Determines the size of the color picker's trigger}

\item{swatches}{One or more predefined color swatches to display as
presets in the color picker. Can include any format the color picker
can parse, including HEX(A), RGB(A), HSL(A), HSV(A), and CSS color
names. Each color must be separated by a semicolon (\verb{;}).
Alternatively, you can pass an array of color values to this property
using JavaScript.}

\item{uppercase}{By default, values are lowercase. With this attribute,
values will be uppercase instead.}

\item{with_hint}{Optional Web Awesome attribute.}

\item{with_label}{Optional Web Awesome attribute.}

\item{without_format_toggle}{Removes the button that lets users toggle
between format.}

\item{hint_slot}{The color picker's form hint. Alternatively, you can
use the \code{hint} attribute.}

\item{label_slot}{The color picker's form label. Alternatively, you can
use the \code{label} attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-color-picker} component.
Generated from Web Awesome metadata.
}
