% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_copy_button.R
\name{wa_copy_button}
\alias{wa_copy_button}
\title{Create a \code{wa-copy-button} component}
\usage{
wa_copy_button(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  value = NULL,
  disabled = NULL,
  copy_label = NULL,
  dir = NULL,
  error_label = NULL,
  feedback_duration = NULL,
  from = NULL,
  lang = NULL,
  success_label = NULL,
  tooltip_placement = NULL,
  copy_icon = NULL,
  error_icon = NULL,
  success_icon = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{String. Default: \code{""}. The text value to copy.}

\item{disabled}{Boolean. Default: \code{FALSE}. Disables the copy button.}

\item{copy_label}{String. Default: \code{""}. A custom label to show in the
tooltip.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{error_label}{String. Default: \code{""}. A custom label to show in the
tooltip when a copy error occurs.}

\item{feedback_duration}{Number. Default: \code{1000}. The length of time to
show feedback before restoring the default trigger.}

\item{from}{String. Default: \code{""}. An id that references an element in
the same document from which data will be copied. If both this and
\code{value} are present, this value will take precedence. By default, the
target element's \code{textContent} will be copied. To copy an attribute,
append the attribute name wrapped in square brackets, e.g.
\code{from="elvalue"}. To copy a property, append a dot and the property
name, e.g. \code{from="el.value"}.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{success_label}{String. Default: \code{""}. A custom label to show in
the tooltip after copying.}

\item{tooltip_placement}{Enumerated string. Allowed values: \code{bottom},
\code{left}, \code{right}, \code{top}. Default: \code{top}. The preferred placement of the
tooltip.}

\item{copy_icon}{The icon to show in the default copy state. Works best
with \verb{<wa-icon>}.}

\item{error_icon}{The icon to show when a copy error occurs. Works best
with \verb{<wa-icon>}.}

\item{success_icon}{The icon to show when the content is copied. Works
best with \verb{<wa-icon>}.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-copy-button} component.
Generated from Web Awesome metadata.
}
