% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_details.R
\name{wa_details}
\alias{wa_details}
\title{Create a \code{wa-details} component}
\usage{
wa_details(
  input_id,
  ...,
  class = NULL,
  style = NULL,
  disabled = NULL,
  name = NULL,
  appearance = NULL,
  dir = NULL,
  icon_placement = NULL,
  lang = NULL,
  open = NULL,
  summary = NULL,
  collapse_icon = NULL,
  expand_icon = NULL,
  summary_slot = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{disabled}{Boolean. Default: \code{FALSE}. Disables the details so it
can't be toggled.}

\item{name}{String. Groups related details elements. When one opens,
others with the same name will close.}

\item{appearance}{Enumerated string. Allowed values: \code{filled},
\code{filled-outlined}, \code{outlined}, \code{plain}. Default: \code{outlined}. The
element's visual appearance.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{icon_placement}{Enumerated string. Allowed values: \code{end},
\code{start}. Default: \code{end}. The location of the expand/collapse icon.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{open}{Boolean. Default: \code{FALSE}. Indicates whether or not the
details is open. You can toggle this attribute to show and hide the
details, or you can use the \code{show()} and \code{hide()} methods and this
attribute will reflect the details' open state.}

\item{summary}{String. The summary to show in the header. If you need
to display HTML, use the \code{summary} slot instead.}

\item{collapse_icon}{Optional collapse icon to use instead of the
default. Works best with \verb{<wa-icon>}.}

\item{expand_icon}{Optional expand icon to use instead of the default.
Works best with \verb{<wa-icon>}.}

\item{summary_slot}{The details' summary. Alternatively, you can use
the \code{summary} attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-details} component. When used
as a Shiny input, \verb{input$<input_id>} reflects the component's current
semantic \code{open} state.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

\verb{input$<input_id>} reflects the component's current semantic \code{open}
state.
}

