% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_dialog.R
\name{wa_dialog}
\alias{wa_dialog}
\title{Create a \code{wa-dialog} component}
\usage{
wa_dialog(
  input_id,
  ...,
  class = NULL,
  style = NULL,
  label = NULL,
  dir = NULL,
  lang = NULL,
  light_dismiss = NULL,
  open = NULL,
  without_header = NULL,
  footer = NULL,
  header_actions = NULL,
  label_slot = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{label}{String. Default: \code{""}. The dialog's label as displayed in
the header. You should always include a relevant label, as it is
required for proper accessibility. If you need to display HTML, use the
\code{label} slot instead.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{light_dismiss}{Boolean. Default: \code{FALSE}. When enabled, the
dialog will be closed when the user clicks outside of it.}

\item{open}{Boolean. Default: \code{FALSE}. Indicates whether or not the
dialog is open. Toggle this attribute to show and hide the dialog.}

\item{without_header}{Boolean. Default: \code{FALSE}. Disables the header.
This will also remove the default close button.}

\item{footer}{The dialog's footer, usually one or more buttons
representing various options.}

\item{header_actions}{Optional actions to add to the header. Works best
with \verb{<wa-button>}.}

\item{label_slot}{The dialog's label. Alternatively, you can use the
\code{label} attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-dialog} component. When used
as a Shiny input, \verb{input$<input_id>} reflects the component's current
semantic \code{open} state.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

\verb{input$<input_id>} reflects the component's current semantic \code{open}
state.
}

