% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_drawer.R
\name{wa_drawer}
\alias{wa_drawer}
\title{Create a \code{wa-drawer} component}
\usage{
wa_drawer(
  input_id,
  ...,
  class = NULL,
  style = NULL,
  label = NULL,
  dir = NULL,
  lang = NULL,
  light_dismiss = NULL,
  open = NULL,
  placement = NULL,
  without_header = NULL,
  footer = NULL,
  header_actions = NULL,
  label_slot = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{label}{The drawer's label as displayed in the header. You should
always include a relevant label, as it is required for proper
accessibility. If you need to display HTML, use the \code{label} slot
instead. Defaults to `` when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{light_dismiss}{When enabled, the drawer will be closed when the
user clicks outside of it. Defaults to \code{true} when omitted.}

\item{open}{Indicates whether or not the drawer is open. Toggle this
attribute to show and hide the drawer. Defaults to \code{false} when
omitted.}

\item{placement}{The direction from which the drawer will open. Must be
one of \code{"bottom"}, \code{"end"}, \code{"start"}, \code{"top"}. Defaults to \code{end} when
omitted.}

\item{without_header}{Disables the header. This will also remove the
default close button. Defaults to \code{false} when omitted.}

\item{footer}{The drawer's footer, usually one or more buttons
representing various options.}

\item{header_actions}{Optional actions to add to the header. Works best
with \verb{<wa-button>}.}

\item{label_slot}{The drawer's label. Alternatively, you can use the
\code{label} attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-drawer} component. When used
as a Shiny input, \verb{input$<input_id>} reflects the component's current
semantic \code{open} state.
Generated from Web Awesome metadata.
}
