% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_dropdown_item.R
\name{wa_dropdown_item}
\alias{wa_dropdown_item}
\title{Create a \code{wa-dropdown-item} component}
\usage{
wa_dropdown_item(
  ...,
  id = NULL,
  value = NULL,
  checked = NULL,
  disabled = NULL,
  dir = NULL,
  lang = NULL,
  submenu_open = NULL,
  type = NULL,
  variant = NULL,
  details = NULL,
  icon = NULL,
  submenu = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{value}{An optional value for the menu item. This is useful for
determining which item was selected when listening to the dropdown's
\code{wa-select} event.}

\item{checked}{Set to true to check the dropdown item. Only valid when
\code{type} is \code{checkbox}. Defaults to \code{false} when omitted.}

\item{disabled}{Disables the dropdown item. Defaults to \code{false} when
omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{submenu_open}{Whether the submenu is currently open. Defaults to
\code{false} when omitted.}

\item{type}{Set to \code{checkbox} to make the item a checkbox. Must be one
of \code{"checkbox"}, \code{"normal"}. Defaults to \code{normal} when omitted.}

\item{variant}{The type of menu item to render. Must be one of
\code{"danger"}, \code{"default"}. Defaults to \code{default} when omitted.}

\item{details}{Additional content or details to display after the
label.}

\item{icon}{An optional icon to display before the label.}

\item{submenu}{Submenu items, typically \verb{<wa-dropdown-item>} elements,
to create a nested menu.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-dropdown-item} component.
Generated from Web Awesome metadata.
}
