% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_format_date.R
\name{wa_format_date}
\alias{wa_format_date}
\title{Create a \code{wa-format-date} component}
\usage{
wa_format_date(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  date = NULL,
  day = NULL,
  dir = NULL,
  era = NULL,
  hour = NULL,
  hour_format = NULL,
  lang = NULL,
  minute = NULL,
  month = NULL,
  second = NULL,
  time_zone = NULL,
  time_zone_name = NULL,
  weekday = NULL,
  year = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{date}{Default: \verb{new Date()}. The date/time to format. If not set,
the current date and time will be used. When passing a string, it's
strongly recommended to use the ISO 8601 format to ensure timezones are
handled correctly. To convert a date to this format in JavaScript, use
\code{date.toISOString()}.}

\item{day}{Enumerated string. Allowed values: \code{2-digit}, \code{numeric}. The
format for displaying the day.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{era}{Enumerated string. Allowed values: \code{long}, \code{narrow},
\code{short}. The format for displaying the era.}

\item{hour}{Enumerated string. Allowed values: \code{2-digit}, \code{numeric}.
The format for displaying the hour.}

\item{hour_format}{Enumerated string. Allowed values: \code{12}, \code{24},
\code{auto}. Default: \code{auto}. The format for displaying the hour.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{minute}{Enumerated string. Allowed values: \code{2-digit}, \code{numeric}.
The format for displaying the minute.}

\item{month}{Enumerated string. Allowed values: \code{2-digit}, \code{long},
\code{narrow}, \code{numeric}, \code{short}. The format for displaying the month.}

\item{second}{Enumerated string. Allowed values: \code{2-digit}, \code{numeric}.
The format for displaying the second.}

\item{time_zone}{String. The time zone to express the time in.}

\item{time_zone_name}{Enumerated string. Allowed values: \code{long},
\code{short}. The format for displaying the time.}

\item{weekday}{Enumerated string. Allowed values: \code{long}, \code{narrow},
\code{short}. The format for displaying the weekday.}

\item{year}{Enumerated string. Allowed values: \code{2-digit}, \code{numeric}.
The format for displaying the year.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-format-date} component.
Generated from Web Awesome metadata.
}
