% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_format_number.R
\name{wa_format_number}
\alias{wa_format_number}
\title{Create a \code{wa-format-number} component}
\usage{
wa_format_number(
  ...,
  id = NULL,
  value = NULL,
  currency = NULL,
  currency_display = NULL,
  dir = NULL,
  lang = NULL,
  maximum_fraction_digits = NULL,
  maximum_significant_digits = NULL,
  minimum_fraction_digits = NULL,
  minimum_integer_digits = NULL,
  minimum_significant_digits = NULL,
  type = NULL,
  without_grouping = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{value}{The number to format. Defaults to \code{0} when omitted.}

\item{currency}{The ISO 4217 currency code to use when formatting.
Defaults to \code{USD} when omitted.}

\item{currency_display}{How to display the currency. Must be one of
\code{"code"}, \code{"name"}, \code{"narrowSymbol"}, \code{"symbol"}. Defaults to \code{symbol}
when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{maximum_fraction_digits}{The maximum number of fraction digits to
use. Possible values are 0-100.}

\item{maximum_significant_digits}{The maximum number of significant
digits to use,. Possible values are 1-21.}

\item{minimum_fraction_digits}{The minimum number of fraction digits to
use. Possible values are 0-100.}

\item{minimum_integer_digits}{The minimum number of integer digits to
use. Possible values are 1-21.}

\item{minimum_significant_digits}{The minimum number of significant
digits to use. Possible values are 1-21.}

\item{type}{The formatting style to use. Must be one of \code{"currency"},
\code{"decimal"}, \code{"percent"}. Defaults to \code{decimal} when omitted.}

\item{without_grouping}{Turns off grouping separators. Defaults to
\code{false} when omitted.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-format-number} component.
Generated from Web Awesome metadata.
}
