% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_icon.R
\name{wa_icon}
\alias{wa_icon}
\title{Create a \code{wa-icon} component}
\usage{
wa_icon(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  label = NULL,
  name = NULL,
  animation = NULL,
  auto_width = NULL,
  dir = NULL,
  family = NULL,
  flip = NULL,
  lang = NULL,
  library = NULL,
  rotate = NULL,
  src = NULL,
  swap_opacity = NULL,
  variant = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{label}{An alternate description to use for assistive devices. If
omitted, the icon will be considered presentational and ignored by
assistive devices. Defaults to `` when omitted.}

\item{name}{The name of the icon to draw. Available names depend on the
icon library being used.}

\item{animation}{Sets the animation for the icon}

\item{auto_width}{Sets the width of the icon to match the cropped SVG
viewBox. This operates like the Font \code{fa-width-auto} class. Defaults to
\code{false} when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{family}{The family of icons to choose from. For Font Awesome
Free, valid options include \code{classic} and \code{brands}. For Font Awesome
Pro subscribers, valid options include, \code{classic}, \code{sharp}, \code{duotone},
\code{sharp-duotone}, and \code{brands}. A valid kit code must be present to show
pro icons via CDN. You can set \verb{<html data-fa-kit-code="...">} to
provide one.}

\item{flip}{Sets the flip direction of the icon along the 'x'
(horizontal), 'y' (vertical), or 'both' axes.}

\item{lang}{Optional Web Awesome attribute.}

\item{library}{The name of a registered custom icon library. Defaults
to \code{default} when omitted.}

\item{rotate}{Sets the rotation degree of the icon Defaults to \code{0} when
omitted.}

\item{src}{An external URL of an SVG file. Be sure you trust the
content you are including, as it will be executed as code and can
result in XSS attacks.}

\item{swap_opacity}{Swaps the opacity of duotone icons. Defaults to
\code{false} when omitted.}

\item{variant}{The name of the icon's variant. For Font Awesome, valid
options include \code{thin}, \code{light}, \code{regular}, and \code{solid} for the
\code{classic} and \code{sharp} families. Some variants require a Font Awesome
Pro subscription. Custom icon libraries may or may not use this
property.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-icon} component.
Generated from Web Awesome metadata.
}
