% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_markdown.R
\name{wa_markdown}
\alias{wa_markdown}
\title{Create a \code{wa-markdown} component}
\usage{
wa_markdown(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  dir = NULL,
  lang = NULL,
  tab_size = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{tab_size}{Number. Default: \code{4}. The tab stop width used when
converting leading tabs to spaces during whitespace normalization.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-markdown} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

