% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_mutation_observer.R
\name{wa_mutation_observer}
\alias{wa_mutation_observer}
\title{Create a \code{wa-mutation-observer} component}
\usage{
wa_mutation_observer(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  disabled = NULL,
  attr = NULL,
  attr_old_value = NULL,
  char_data = NULL,
  char_data_old_value = NULL,
  child_list = NULL,
  dir = NULL,
  lang = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{disabled}{Boolean. Default: \code{FALSE}. Disables the observer.}

\item{attr}{String. Watches for changes to attributes. To watch only
specific attributes, separate them by a space, e.g. \code{attr="class id title"}. To watch all attributes, use \code{*}.}

\item{attr_old_value}{Boolean. Default: \code{FALSE}. Indicates whether or
not the attribute's previous value should be recorded when monitoring
changes.}

\item{char_data}{Boolean. Default: \code{FALSE}. Watches for changes to the
character data contained within the node.}

\item{char_data_old_value}{Boolean. Default: \code{FALSE}. Indicates whether
or not the previous value of the node's text should be recorded.}

\item{child_list}{Boolean. Default: \code{FALSE}. Watches for the addition
or removal of new child nodes.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-mutation-observer} component.
Generated from Web Awesome metadata.
}
