% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_option.R
\name{wa_option}
\alias{wa_option}
\title{Create a \code{wa-option} component}
\usage{
wa_option(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  value = NULL,
  disabled = NULL,
  label = NULL,
  dir = NULL,
  lang = NULL,
  selected = NULL,
  end = NULL,
  start = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{String. Default: \code{""}. The option's value. When selected,
the containing form control will receive this value. The value must be
unique from other options in the same group. Values may not contain
spaces, as spaces are used as delimiters when listing multiple values.}

\item{disabled}{Boolean. Default: \code{FALSE}. Draws the option in a
disabled state, preventing selection.}

\item{label}{String. The option’s plain text label. Usually
automatically generated, but can be useful to provide manually for
cases involving complex content.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{selected}{Boolean. Default: \code{FALSE}. Selects an option initially.
This wrapper argument sets the HTML \code{selected} attribute, which maps to
the component's \code{defaultSelected} field/property.}

\item{end}{An element, such as \verb{<wa-icon>}, placed after the label.}

\item{start}{An element, such as \verb{<wa-icon>}, placed before the label.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-option} component.
Generated from Web Awesome metadata.
}
