% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_progress_bar.R
\name{wa_progress_bar}
\alias{wa_progress_bar}
\title{Create a \code{wa-progress-bar} component}
\usage{
wa_progress_bar(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  value = NULL,
  label = NULL,
  dir = NULL,
  indeterminate = NULL,
  lang = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{Number. Default: \code{0}. The current progress as a
percentage, 0 to 100.}

\item{label}{String. Default: \code{""}. A custom label for assistive
devices.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{indeterminate}{Boolean. Default: \code{FALSE}. When TRUE, percentage
is ignored, the label is hidden, and the progress bar is drawn in an
indeterminate state.}

\item{lang}{String. Optional Web Awesome attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-progress-bar} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

