% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_progress_ring.R
\name{wa_progress_ring}
\alias{wa_progress_ring}
\title{Create a \code{wa-progress-ring} component}
\usage{
wa_progress_ring(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  value = NULL,
  label = NULL,
  dir = NULL,
  lang = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{Number. Default: \code{0}. The current progress as a
percentage, 0 to 100.}

\item{label}{String. Default: \code{""}. A custom label for assistive
devices.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-progress-ring} component.
Generated from Web Awesome metadata.
}
