% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_qr_code.R
\name{wa_qr_code}
\alias{wa_qr_code}
\title{Create a \code{wa-qr-code} component}
\usage{
wa_qr_code(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  value = NULL,
  label = NULL,
  background = NULL,
  dir = NULL,
  error_correction = NULL,
  fill = NULL,
  lang = NULL,
  radius = NULL,
  size = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{String. Default: \code{""}. The QR code's value.}

\item{label}{String. Default: \code{""}. The label for assistive devices to
announce. If unspecified, the value will be used instead.}

\item{background}{String. Default: \code{""}. The background color. This can
be any valid CSS color or \code{transparent}. It cannot be a CSS custom
property.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{error_correction}{Enumerated string. Allowed values: \code{H}, \code{L},
\code{M}, \code{Q}. Default: \code{H}. The level of error correction to use. Learn
more}

\item{fill}{String. Default: \code{""}. The fill color. This can be any
valid CSS color, but not a CSS custom property.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{radius}{Number. Default: \code{0}. The edge radius of each module.
Must be between 0 and 0.5.}

\item{size}{Number. Default: \code{128}. The size of the QR code, in pixels.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-qr-code} component.
Generated from Web Awesome metadata.
}
