% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_radio.R
\name{wa_radio}
\alias{wa_radio}
\title{Create a \code{wa-radio} component}
\usage{
wa_radio(
  ...,
  id = NULL,
  value = NULL,
  disabled = NULL,
  name = NULL,
  appearance = NULL,
  custom_error = NULL,
  dir = NULL,
  lang = NULL,
  size = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{value}{The radio's value. When selected, the radio group will
receive this value.}

\item{disabled}{Disables the radio. Defaults to \code{false} when omitted.}

\item{name}{The name of the input, submitted as a name/value pair with
form data. Defaults to \code{null} when omitted.}

\item{appearance}{The radio's visual appearance. Must be one of
\code{"button"}, \code{"default"}. Defaults to \code{default} when omitted.}

\item{custom_error}{Optional Web Awesome attribute. Defaults to \code{null}
when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{size}{The radio's size. When used inside a radio group, the size
will be determined by the radio group's size, which will override this
attribute. Must be one of \code{"large"}, \code{"medium"}, \code{"small"}.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-radio} component.
Generated from Web Awesome metadata.
}
