% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_rating.R
\name{wa_rating}
\alias{wa_rating}
\title{Create a \code{wa-rating} component}
\usage{
wa_rating(
  input_id,
  ...,
  class = NULL,
  style = NULL,
  value = NULL,
  disabled = NULL,
  label = NULL,
  dir = NULL,
  get_symbol = NULL,
  lang = NULL,
  max = NULL,
  precision = NULL,
  readonly = NULL,
  size = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{Number. Default: \code{0}. The current rating.}

\item{disabled}{Boolean. Default: \code{FALSE}. Disables the rating.}

\item{label}{String. Default: \code{""}. A label that describes the rating
to assistive devices.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{get_symbol}{A function that customizes the symbol to be rendered.
The first and only argument is the rating's current value. The function
should return a string containing trusted HTML of the symbol to render
at the specified value. Works well with \verb{<wa-icon>} elements.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{max}{Number. Default: \code{5}. The highest rating to show.}

\item{precision}{Number. Default: \code{1}. The precision at which the
rating will increase and decrease. For example, to allow half-star
ratings, set this attribute to \code{0.5}.}

\item{readonly}{Boolean. Default: \code{FALSE}. Makes the rating readonly.}

\item{size}{Enumerated string. Allowed values: \code{large}, \code{medium},
\code{small}. Default: \code{medium}. The component's size.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-rating} component. When used
as a Shiny input, \verb{input$<input_id>} reflects the component's current
\code{value} value.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

\verb{input$<input_id>} reflects the component's current \code{value} value.
}

