% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_relative_time.R
\name{wa_relative_time}
\alias{wa_relative_time}
\title{Create a \code{wa-relative-time} component}
\usage{
wa_relative_time(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  date = NULL,
  dir = NULL,
  format = NULL,
  lang = NULL,
  numeric = NULL,
  sync = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{date}{The date from which to calculate time from. If not set, the
current date and time will be used. When passing a string, it's
strongly recommended to use the ISO 8601 format to ensure timezones are
handled correctly. To convert a date to this format in JavaScript, use
\code{date.toISOString()}. Defaults to \verb{new Date()} when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{format}{The formatting style to use. Must be one of \code{"long"},
\code{"narrow"}, \code{"short"}. Defaults to \code{long} when omitted.}

\item{lang}{Optional Web Awesome attribute.}

\item{numeric}{When \code{auto}, values such as "yesterday" and "tomorrow"
will be shown when possible. When \code{always}, values such as "1 day ago"
and "in 1 day" will be shown. Must be one of \code{"always"}, \code{"auto"}.
Defaults to \code{auto} when omitted.}

\item{sync}{Keep the displayed value up to date as time passes.
Defaults to \code{false} when omitted.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-relative-time} component.
Generated from Web Awesome metadata.
}
