% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_resize_observer.R
\name{wa_resize_observer}
\alias{wa_resize_observer}
\title{Create a \code{wa-resize-observer} component}
\usage{
wa_resize_observer(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  disabled = NULL,
  dir = NULL,
  lang = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{disabled}{Disables the observer. Defaults to \code{false} when
omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-resize-observer} component.
Generated from Web Awesome metadata.
}
