% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_scroller.R
\name{wa_scroller}
\alias{wa_scroller}
\title{Create a \code{wa-scroller} component}
\usage{
wa_scroller(
  ...,
  id = NULL,
  dir = NULL,
  lang = NULL,
  orientation = NULL,
  without_scrollbar = NULL,
  without_shadow = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{orientation}{The scroller's orientation. Must be one of
\code{"horizontal"}, \code{"vertical"}. Defaults to \code{horizontal} when omitted.}

\item{without_scrollbar}{Removes the visible scrollbar. Defaults to
\code{false} when omitted.}

\item{without_shadow}{Removes the shadows. Defaults to \code{false} when
omitted.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-scroller} component.
Generated from Web Awesome metadata.
}
