% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_skeleton.R
\name{wa_skeleton}
\alias{wa_skeleton}
\title{Create a \code{wa-skeleton} component}
\usage{
wa_skeleton(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  dir = NULL,
  effect = NULL,
  lang = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{dir}{Optional Web Awesome attribute.}

\item{effect}{Determines which effect the skeleton will use. Must be
one of \code{"none"}, \code{"pulse"}, \code{"sheen"}. Defaults to \code{none} when omitted.}

\item{lang}{Optional Web Awesome attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-skeleton} component.
Generated from Web Awesome metadata.
}
