% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_split_panel.R
\name{wa_split_panel}
\alias{wa_split_panel}
\title{Create a \code{wa-split-panel} component}
\usage{
wa_split_panel(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  disabled = NULL,
  dir = NULL,
  lang = NULL,
  orientation = NULL,
  position = NULL,
  position_in_pixels = NULL,
  primary = NULL,
  snap = NULL,
  snap_threshold = NULL,
  divider = NULL,
  end = NULL,
  start = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{disabled}{Disables resizing. Note that the position may still
change as a result of resizing the host element. Defaults to \code{false}
when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{orientation}{Sets the split panel's orientation. Must be one of
\code{"horizontal"}, \code{"vertical"}. Defaults to \code{horizontal} when omitted.}

\item{position}{The current position of the divider from the primary
panel's edge as a percentage 0-100. Defaults to 50\% of the container's
initial size. Defaults to \code{50} when omitted.}

\item{position_in_pixels}{The current position of the divider from the
primary panel's edge in pixels.}

\item{primary}{If no primary panel is designated, both panels will
resize proportionally when the host element is resized. If a primary
panel is designated, it will maintain its size and the other panel will
grow or shrink as needed when the host element is resized.}

\item{snap}{One or more space-separated values at which the divider
should snap. Values can be in pixels or percentages, e.g. \code{"100px 50\%"}.}

\item{snap_threshold}{How close the divider must be to a snap point
until snapping occurs. Defaults to \code{12} when omitted.}

\item{divider}{The divider. Useful for slotting in a custom icon that
renders as a handle.}

\item{end}{Content to place in the end panel.}

\item{start}{Content to place in the start panel.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-split-panel} component.
Generated from Web Awesome metadata.
}
