% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_switch.R
\name{wa_switch}
\alias{wa_switch}
\title{Create a \code{wa-switch} component}
\usage{
wa_switch(
  input_id,
  ...,
  value = NULL,
  checked = NULL,
  disabled = NULL,
  hint = NULL,
  name = NULL,
  custom_error = NULL,
  dir = NULL,
  lang = NULL,
  required = NULL,
  size = NULL,
  title = NULL,
  with_hint = NULL,
  hint_slot = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{value}{The value of the switch, submitted as a name/value pair
with form data.}

\item{checked}{The default value of the form control. Primarily used
for resetting the form control.}

\item{disabled}{Disables the switch. Defaults to \code{false} when omitted.}

\item{hint}{The switch's hint. If you need to display HTML, use the
\code{hint} slot instead. Defaults to `` when omitted.}

\item{name}{The name of the switch, submitted as a name/value pair with
form data. Defaults to \code{null} when omitted.}

\item{custom_error}{Optional Web Awesome attribute. Defaults to \code{null}
when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{required}{Makes the switch a required field. Defaults to \code{false}
when omitted.}

\item{size}{The switch's size. Must be one of \code{"large"}, \code{"medium"},
\code{"small"}. Defaults to \code{medium} when omitted.}

\item{title}{Optional Web Awesome attribute. Defaults to `` when
omitted.}

\item{with_hint}{Used for SSR. If you slot in hint, make sure to add
\code{with-hint} to your component to get it to properly render with SSR.
Defaults to \code{false} when omitted.}

\item{hint_slot}{Text that describes how to use the switch.
Alternatively, you can use the \code{hint} attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-switch} component.
Generated from Web Awesome metadata.
}
