% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_tab_group.R
\name{wa_tab_group}
\alias{wa_tab_group}
\title{Create a \code{wa-tab-group} component}
\usage{
wa_tab_group(
  input_id,
  ...,
  class = NULL,
  style = NULL,
  activation = NULL,
  active = NULL,
  dir = NULL,
  lang = NULL,
  placement = NULL,
  without_scroll_controls = NULL,
  nav = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{activation}{Enumerated string. Allowed values: \code{auto}, \code{manual}.
Default: \code{auto}. When set to auto, navigating tabs with the arrow keys
will instantly show the corresponding tab panel. When set to manual,
the tab will receive focus but will not show until the user presses
spacebar or enter.}

\item{active}{String. Default: \code{""}. Sets the active tab.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{placement}{Enumerated string. Allowed values: \code{bottom}, \code{end},
\code{start}, \code{top}. Default: \code{top}. The placement of the tabs.}

\item{without_scroll_controls}{Boolean. Default: \code{FALSE}. Disables the
scroll arrows that appear when tabs overflow.}

\item{nav}{Used for grouping tabs in the tab group. Must be \verb{<wa-tab>}
elements. Note that \verb{<wa-tab>} will set this slot on itself
automatically.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-tab-group} component. When
used as a Shiny input, \verb{input$<input_id>} reflects the component's
current semantic \code{active} state.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

\verb{input$<input_id>} reflects the component's current semantic \code{active}
state.
}

