% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_tab_panel.R
\name{wa_tab_panel}
\alias{wa_tab_panel}
\title{Create a \code{wa-tab-panel} component}
\usage{
wa_tab_panel(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  name = NULL,
  active = NULL,
  dir = NULL,
  lang = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{name}{String. Default: \code{""}. The tab panel's name.}

\item{active}{Boolean. Default: \code{FALSE}. When TRUE, the tab panel will
be shown.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-tab-panel} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

