% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_tag.R
\name{wa_tag}
\alias{wa_tag}
\title{Create a \code{wa-tag} component}
\usage{
wa_tag(
  ...,
  id = NULL,
  appearance = NULL,
  dir = NULL,
  lang = NULL,
  pill = NULL,
  size = NULL,
  variant = NULL,
  with_remove = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{appearance}{The tag's visual appearance.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{pill}{Draws a pill-style tag with rounded edges.}

\item{size}{The tag's size.}

\item{variant}{The tag's theme variant. Defaults to \code{neutral} if not
within another element with a variant.}

\item{with_remove}{Makes the tag removable and shows a remove button.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-tag} component.
Generated from Web Awesome metadata.
}
