% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_tooltip.R
\name{wa_tooltip}
\alias{wa_tooltip}
\title{Create a \code{wa-tooltip} component}
\usage{
wa_tooltip(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  disabled = NULL,
  dir = NULL,
  distance = NULL,
  `for` = NULL,
  hide_delay = NULL,
  lang = NULL,
  open = NULL,
  placement = NULL,
  show_delay = NULL,
  skidding = NULL,
  trigger = NULL,
  without_arrow = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{disabled}{Boolean. Default: \code{FALSE}. Disables the tooltip so it
won't show when triggered.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{distance}{Number. Default: \code{8}. The distance in pixels from which
to offset the tooltip away from its target.}

\item{for}{String. Default: \code{null}. Optional Web Awesome attribute.}

\item{hide_delay}{Number. Default: \code{0}. The amount of time to wait
before hiding the tooltip when the user mouses out.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{open}{Boolean. Default: \code{FALSE}. Indicates whether or not the
tooltip is open. You can use this in lieu of the show/hide methods.}

\item{placement}{Enumerated string. Allowed values: \code{bottom},
\code{bottom-end}, \code{bottom-start}, \code{left}, \code{left-end}, \code{left-start},
\code{right}, \code{right-end}, \code{right-start}, \code{top}, \code{top-end}, \code{top-start}.
Default: \code{top}. The preferred placement of the tooltip. Note that the
actual placement may vary as needed to keep the tooltip inside of the
viewport.}

\item{show_delay}{Number. Default: \code{150}. The amount of time to wait
before showing the tooltip when the user mouses in.}

\item{skidding}{Number. Default: \code{0}. The distance in pixels from which
to offset the tooltip along its target.}

\item{trigger}{String. Default: \verb{hover focus}. Controls how the tooltip
is activated. Possible options include \code{click}, \code{hover}, \code{focus}, and
\code{manual}. Multiple options can be passed by separating them with a
space. When manual is used, the tooltip must be activated
programmatically.}

\item{without_arrow}{Boolean. Default: \code{FALSE}. Removes the arrow from
the tooltip.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-tooltip} component.
Generated from Web Awesome metadata.
}
