% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_tree_item.R
\name{wa_tree_item}
\alias{wa_tree_item}
\title{Create a \code{wa-tree-item} component}
\usage{
wa_tree_item(
  ...,
  id = NULL,
  disabled = NULL,
  dir = NULL,
  expanded = NULL,
  lang = NULL,
  lazy = NULL,
  selected = NULL,
  collapse_icon = NULL,
  expand_icon = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{disabled}{Disables the tree item. Defaults to \code{false} when
omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{expanded}{Expands the tree item. Defaults to \code{false} when
omitted.}

\item{lang}{Optional Web Awesome attribute.}

\item{lazy}{Enables lazy loading behavior. Defaults to \code{false} when
omitted.}

\item{selected}{Draws the tree item in a selected state. Defaults to
\code{false} when omitted.}

\item{collapse_icon}{The icon to show when the tree item is collapsed.}

\item{expand_icon}{The icon to show when the tree item is expanded.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-tree-item} component.
Generated from Web Awesome metadata.
}
