% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_version.R
\name{wa_version}
\alias{wa_version}
\title{Return the bundled Web Awesome version}
\usage{
wa_version()
}
\value{
A length-1 character vector containing the bundled Web Awesome
version string.
}
\description{
Reports the Web Awesome version bundled with the current
\code{shiny.webawesome} installation.
}
\examples{
wa_version()
}
