% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_zoomable_frame.R
\name{wa_zoomable_frame}
\alias{wa_zoomable_frame}
\title{Create a \code{wa-zoomable-frame} component}
\usage{
wa_zoomable_frame(
  ...,
  id = NULL,
  allowfullscreen = NULL,
  dir = NULL,
  lang = NULL,
  loading = NULL,
  referrerpolicy = NULL,
  sandbox = NULL,
  src = NULL,
  srcdoc = NULL,
  without_controls = NULL,
  without_interaction = NULL,
  zoom = NULL,
  zoom_levels = NULL,
  zoom_in_icon = NULL,
  zoom_out_icon = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{allowfullscreen}{Allows fullscreen mode. Defaults to \code{false} when
omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{loading}{Controls iframe loading behavior. Must be one of
\code{"eager"}, \code{"lazy"}. Defaults to \code{eager} when omitted.}

\item{referrerpolicy}{Controls referrer information.}

\item{sandbox}{Security restrictions for the iframe.}

\item{src}{The URL of the content to display.}

\item{srcdoc}{Inline HTML to display.}

\item{without_controls}{Removes the zoom controls. Defaults to \code{false}
when omitted.}

\item{without_interaction}{Disables interaction when present. Defaults
to \code{false} when omitted.}

\item{zoom}{The current zoom of the frame, e.g. 0 = 0\% and 1 = 100\%.
Defaults to \code{1} when omitted.}

\item{zoom_levels}{The zoom levels to step through when using zoom
controls. This does not restrict programmatic changes to the zoom.
Defaults to \verb{25\% 50\% 75\% 100\% 125\% 150\% 175\% 200\%} when omitted.}

\item{zoom_in_icon}{The slot that contains the zoom in icon.}

\item{zoom_out_icon}{The slot that contains the zoom out icon.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-zoomable-frame} component.
Generated from Web Awesome metadata.
}
