% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_animation.R
\name{wa_animation}
\alias{wa_animation}
\title{Create a \code{wa-animation} component}
\usage{
wa_animation(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  name = NULL,
  delay = NULL,
  dir = NULL,
  direction = NULL,
  duration = NULL,
  easing = NULL,
  end_delay = NULL,
  fill = NULL,
  iteration_start = NULL,
  iterations = NULL,
  lang = NULL,
  play = NULL,
  playback_rate = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{name}{String. Default: \code{none}. The name of the built-in animation
to use. For custom animations, use the \code{keyframes} prop.}

\item{delay}{Number. Default: \code{0}. The number of milliseconds to delay
the start of the animation.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{direction}{Default: \code{'normal'}. Determines the direction of
playback as well as the behavior when reaching the end of an iteration.
Learn more}

\item{duration}{Number. Default: \code{1000}. The number of milliseconds
each iteration of the animation takes to complete.}

\item{easing}{String. Default: \code{linear}. The easing function to use for
the animation. This can be a Web Awesome easing function or a custom
easing function such as \code{cubic-bezier(0, 1, .76, 1.14)}.}

\item{end_delay}{Number. Default: \code{0}. The number of milliseconds to
delay after the active period of an animation sequence.}

\item{fill}{Default: \code{'auto'}. Sets how the animation applies styles to
its target before and after its execution.}

\item{iteration_start}{Number. Default: \code{0}. The offset at which to
start the animation, usually between 0 (start) and 1 (end).}

\item{iterations}{Number. Default: \code{Infinity}. The number of iterations
to run before the animation completes. Defaults to \code{Infinity}, which
loops.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{play}{Boolean. Default: \code{FALSE}. Plays the animation. When
omitted, the animation will be paused. This attribute will be
automatically removed when the animation finishes or gets canceled.}

\item{playback_rate}{Number. Default: \code{1}. Sets the animation's
playback rate. The default is \code{1}, which plays the animation at a
normal speed. Setting this to \code{2}, for example, will double the
animation's speed. A negative value can be used to reverse the
animation. This value can be changed without causing the animation to
restart.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-animation} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

