% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_avatar.R
\name{wa_avatar}
\alias{wa_avatar}
\title{Create a \code{wa-avatar} component}
\usage{
wa_avatar(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  label = NULL,
  dir = NULL,
  image = NULL,
  initials = NULL,
  lang = NULL,
  loading = NULL,
  shape = NULL,
  icon = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{label}{String. Default: \code{""}. A label to use to describe the
avatar to assistive devices.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{image}{String. Default: \code{""}. The image source to use for the
avatar.}

\item{initials}{String. Default: \code{""}. Initials to use as a fallback
when no image is available (1-2 characters max recommended).}

\item{lang}{String. Optional Web Awesome attribute.}

\item{loading}{Enumerated string. Allowed values: \code{eager}, \code{lazy}.
Default: \code{eager}. Indicates how the browser should load the image.}

\item{shape}{Enumerated string. Allowed values: \code{circle}, \code{rounded},
\code{square}. Default: \code{circle}. The shape of the avatar.}

\item{icon}{The default icon to use when no image or initials are
present. Works best with \verb{<wa-icon>}.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-avatar} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

