% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_button.R
\name{wa_button}
\alias{wa_button}
\title{Create a \code{wa-button} component}
\usage{
wa_button(
  input_id,
  ...,
  class = NULL,
  style = NULL,
  value = NULL,
  disabled = NULL,
  name = NULL,
  appearance = NULL,
  custom_error = NULL,
  dir = NULL,
  download = NULL,
  formaction = NULL,
  formenctype = NULL,
  formmethod = NULL,
  formnovalidate = NULL,
  formtarget = NULL,
  href = NULL,
  lang = NULL,
  loading = NULL,
  pill = NULL,
  rel = NULL,
  size = NULL,
  target = NULL,
  title = NULL,
  type = NULL,
  variant = NULL,
  with_caret = NULL,
  end = NULL,
  start = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{The value of the button, submitted as a pair with the
button's name as part of the form data, but only when this button is
the submitter. This attribute is ignored when \code{href} is present.}

\item{disabled}{Disables the button. Defaults to \code{false} when omitted.}

\item{name}{The name of the button, submitted as a name/value pair with
form data, but only when this button is the submitter. This attribute
is ignored when \code{href} is present. Defaults to \code{null} when omitted.}

\item{appearance}{The button's visual appearance. Must be one of
\code{"accent"}, \code{"filled"}, \code{"filled-outlined"}, \code{"outlined"}, \code{"plain"}.
Defaults to \code{accent} when omitted.}

\item{custom_error}{Optional Web Awesome attribute. Defaults to \code{null}
when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{download}{Tells the browser to download the linked file as this
filename. Only used when \code{href} is present.}

\item{formaction}{Used to override the form owner's \code{action} attribute.
This wrapper argument sets the HTML \code{formaction} attribute, which maps
to the component's \code{formAction} field/property.}

\item{formenctype}{Used to override the form owner's \code{enctype}
attribute. This wrapper argument sets the HTML \code{formenctype} attribute,
which maps to the component's \code{formEnctype} field/property. Must be one
of \code{"application/x-www-form-urlencoded"}, \code{"multipart/form-data"},
\code{"text/plain"}.}

\item{formmethod}{Used to override the form owner's \code{method} attribute.
This wrapper argument sets the HTML \code{formmethod} attribute, which maps
to the component's \code{formMethod} field/property. Must be one of \code{"get"},
\code{"post"}.}

\item{formnovalidate}{Used to override the form owner's \code{novalidate}
attribute. This wrapper argument sets the HTML \code{formnovalidate}
attribute, which maps to the component's \code{formNoValidate}
field/property.}

\item{formtarget}{Used to override the form owner's \code{target} attribute.
This wrapper argument sets the HTML \code{formtarget} attribute, which maps
to the component's \code{formTarget} field/property.}

\item{href}{When set, the underlying button will be rendered as an
\verb{<a>} with this \code{href} instead of a \verb{<button>}.}

\item{lang}{Optional Web Awesome attribute.}

\item{loading}{Draws the button in a loading state. Defaults to \code{false}
when omitted.}

\item{pill}{Draws a pill-style button with rounded edges. Defaults to
\code{false} when omitted.}

\item{rel}{When using \code{href}, this attribute will map to the underlying
link's \code{rel} attribute.}

\item{size}{The button's size. Must be one of \code{"large"}, \code{"medium"},
\code{"small"}. Defaults to \code{medium} when omitted.}

\item{target}{Tells the browser where to open the link. Only used when
\code{href} is present. Must be one of \code{"_blank"}, \code{"_parent"}, \code{"_self"},
\code{"_top"}.}

\item{title}{Optional Web Awesome attribute. Defaults to `` when
omitted.}

\item{type}{The type of button. Note that the default value is \code{button}
instead of \code{submit}, which is opposite of how native \verb{<button>}
elements behave. When the type is \code{submit}, the button will submit the
surrounding form. Must be one of \code{"button"}, \code{"reset"}, \code{"submit"}.
Defaults to \code{button} when omitted.}

\item{variant}{The button's theme variant. Defaults to \code{neutral} if not
within another element with a variant. Must be one of \code{"brand"},
\code{"danger"}, \code{"neutral"}, \code{"success"}, \code{"warning"}. Defaults to
\code{neutral} when omitted.}

\item{with_caret}{Draws the button with a caret. Used to indicate that
the button triggers a dropdown menu or similar behavior. Defaults to
\code{false} when omitted.}

\item{end}{An element, such as \verb{<wa-icon>}, placed after the label.}

\item{start}{An element, such as \verb{<wa-icon>}, placed before the label.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-button} component.
Generated from Web Awesome metadata.
}
