% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_callout.R
\name{wa_callout}
\alias{wa_callout}
\title{Create a \code{wa-callout} component}
\usage{
wa_callout(
  ...,
  id = NULL,
  appearance = NULL,
  dir = NULL,
  lang = NULL,
  size = NULL,
  variant = NULL,
  icon = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{appearance}{The callout's visual appearance.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{size}{The callout's size.}

\item{variant}{The callout's theme variant. Defaults to \code{brand} if not
within another element with a variant.}

\item{icon}{An icon to show in the callout. Works best with
\verb{<wa-icon>}.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-callout} component.
Generated from Web Awesome metadata.
}
