% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_card.R
\name{wa_card}
\alias{wa_card}
\title{Create a \code{wa-card} component}
\usage{
wa_card(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  appearance = NULL,
  dir = NULL,
  lang = NULL,
  orientation = NULL,
  actions = NULL,
  footer = NULL,
  footer_actions = NULL,
  header = NULL,
  header_actions = NULL,
  media = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{appearance}{The card's visual appearance. Must be one of
\code{"accent"}, \code{"filled"}, \code{"filled-outlined"}, \code{"outlined"}, \code{"plain"}.
Defaults to \code{outlined} when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{orientation}{Renders the card's orientation * Must be one of
\code{"horizontal"}, \code{"vertical"}. Defaults to \code{vertical} when omitted.}

\item{actions}{An optional actions section to render at the end for the
horizontal card.}

\item{footer}{An optional footer for the card.}

\item{footer_actions}{An optional actions section to render in the
footer of the vertical card.}

\item{header}{An optional header for the card.}

\item{header_actions}{An optional actions section to render in the
header of the vertical card.}

\item{media}{An optional media section to render at the start of the
card.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-card} component.
Generated from Web Awesome metadata.
}
