% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_card.R
\name{wa_card}
\alias{wa_card}
\title{Create a \code{wa-card} component}
\usage{
wa_card(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  appearance = NULL,
  dir = NULL,
  lang = NULL,
  orientation = NULL,
  with_footer = NULL,
  with_header = NULL,
  with_media = NULL,
  actions = NULL,
  footer = NULL,
  footer_actions = NULL,
  header = NULL,
  header_actions = NULL,
  media = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{appearance}{Enumerated string. Allowed values: \code{accent},
\code{filled}, \code{filled-outlined}, \code{outlined}, \code{plain}. Default: \code{outlined}.
The card's visual appearance.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{orientation}{Enumerated string. Allowed values: \code{horizontal},
\code{vertical}. Default: \code{vertical}. Renders the card's orientation *}

\item{with_footer}{Boolean. Default: \code{FALSE}. Only required for SSR.
Set to \code{TRUE} if you're slotting in a \code{footer} element so the
server-rendered markup includes the footer before the component
hydrates on the client.}

\item{with_header}{Boolean. Default: \code{FALSE}. Only required for SSR.
Set to \code{TRUE} if you're slotting in a \code{header} element so the
server-rendered markup includes the header before the component
hydrates on the client.}

\item{with_media}{Boolean. Default: \code{FALSE}. Only required for SSR. Set
to \code{TRUE} if you're slotting in a \code{media} element so the
server-rendered markup includes the media before the component hydrates
on the client.}

\item{actions}{An optional actions section to render at the end for the
horizontal card.}

\item{footer}{An optional footer for the card.}

\item{footer_actions}{An optional actions section to render in the
footer of the vertical card.}

\item{header}{An optional header for the card.}

\item{header_actions}{An optional actions section to render in the
header of the vertical card.}

\item{media}{An optional media section to render at the start of the
card.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-card} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

