% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_carousel_item.R
\name{wa_carousel_item}
\alias{wa_carousel_item}
\title{Create a \code{wa-carousel-item} component}
\usage{
wa_carousel_item(..., id = NULL, dir = NULL, lang = NULL)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-carousel-item} component.
Generated from Web Awesome metadata.
}
