% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_checkbox.R
\name{wa_checkbox}
\alias{wa_checkbox}
\title{Create a \code{wa-checkbox} component}
\usage{
wa_checkbox(
  input_id,
  ...,
  value = NULL,
  checked = NULL,
  disabled = NULL,
  hint = NULL,
  name = NULL,
  custom_error = NULL,
  dir = NULL,
  indeterminate = NULL,
  lang = NULL,
  required = NULL,
  size = NULL,
  title = NULL,
  hint_slot = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{value}{The value of the checkbox, submitted as a name/value pair
with form data.}

\item{checked}{The default value of the form control. Primarily used
for resetting the form control.}

\item{disabled}{Disables the checkbox. Defaults to \code{false} when
omitted.}

\item{hint}{The checkbox's hint. If you need to display HTML, use the
\code{hint} slot instead. Defaults to `` when omitted.}

\item{name}{The name of the checkbox, submitted as a name/value pair
with form data. Defaults to \code{null} when omitted.}

\item{custom_error}{Optional Web Awesome attribute. Defaults to \code{null}
when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{indeterminate}{Draws the checkbox in an indeterminate state. This
is usually applied to checkboxes that represents a "select all/none"
behavior when associated checkboxes have a mix of checked and unchecked
states. Defaults to \code{false} when omitted.}

\item{lang}{Optional Web Awesome attribute.}

\item{required}{Makes the checkbox a required field. Defaults to
\code{false} when omitted.}

\item{size}{The checkbox's size. Must be one of \code{"large"}, \code{"medium"},
\code{"small"}. Defaults to \code{medium} when omitted.}

\item{title}{Optional Web Awesome attribute. Defaults to `` when
omitted.}

\item{hint_slot}{Text that describes how to use the checkbox.
Alternatively, you can use the \code{hint} attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-checkbox} component.
Generated from Web Awesome metadata.
}
