% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_comparison.R
\name{wa_comparison}
\alias{wa_comparison}
\title{Create a \code{wa-comparison} component}
\usage{
wa_comparison(
  input_id,
  ...,
  class = NULL,
  style = NULL,
  dir = NULL,
  lang = NULL,
  position = NULL,
  after = NULL,
  before = NULL,
  handle = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{position}{The position of the divider as a percentage. Defaults
to \code{50} when omitted.}

\item{after}{The after content, often an \verb{<img>} or \verb{<svg>} element.}

\item{before}{The before content, often an \verb{<img>} or \verb{<svg>} element.}

\item{handle}{The icon used inside the handle.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-comparison} component.
Generated from Web Awesome metadata.
}
