% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_dialog.R
\name{wa_dialog}
\alias{wa_dialog}
\title{Create a \code{wa-dialog} component}
\usage{
wa_dialog(
  ...,
  id = NULL,
  label = NULL,
  dir = NULL,
  lang = NULL,
  light_dismiss = NULL,
  open = NULL,
  without_header = NULL,
  footer = NULL,
  header_actions = NULL,
  label_slot = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{label}{The dialog's label as displayed in the header. You should
always include a relevant label, as it is required for proper
accessibility. If you need to display HTML, use the \code{label} slot
instead.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{light_dismiss}{When enabled, the dialog will be closed when the
user clicks outside of it.}

\item{open}{Indicates whether or not the dialog is open. Toggle this
attribute to show and hide the dialog.}

\item{without_header}{Disables the header. This will also remove the
default close button.}

\item{footer}{The dialog's footer, usually one or more buttons
representing various options.}

\item{header_actions}{Optional actions to add to the header. Works best
with \verb{<wa-button>}.}

\item{label_slot}{The dialog's label. Alternatively, you can use the
\code{label} attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-dialog} component.
Generated from Web Awesome metadata.
}
