% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_divider.R
\name{wa_divider}
\alias{wa_divider}
\title{Create a \code{wa-divider} component}
\usage{
wa_divider(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  dir = NULL,
  lang = NULL,
  orientation = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{orientation}{Enumerated string. Allowed values: \code{horizontal},
\code{vertical}. Default: \code{horizontal}. Sets the divider's orientation.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-divider} component.
Generated from Web Awesome metadata.
}
