% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_dropdown.R
\name{wa_dropdown}
\alias{wa_dropdown}
\title{Create a \code{wa-dropdown} component}
\usage{
wa_dropdown(
  ...,
  id = NULL,
  dir = NULL,
  distance = NULL,
  lang = NULL,
  open = NULL,
  placement = NULL,
  size = NULL,
  skidding = NULL,
  trigger = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{dir}{Optional Web Awesome attribute.}

\item{distance}{The distance of the dropdown menu from its trigger.
Defaults to \code{0} when omitted.}

\item{lang}{Optional Web Awesome attribute.}

\item{open}{Opens or closes the dropdown. Defaults to \code{false} when
omitted.}

\item{placement}{The placement of the dropdown menu in reference to the
trigger. The menu will shift to a more optimal location if the
preferred placement doesn't have enough room. Must be one of
\code{"bottom"}, \code{"bottom-end"}, \code{"bottom-start"}, \code{"left"}, \code{"left-end"},
\code{"left-start"}, \code{"right"}, \code{"right-end"}, \code{"right-start"}, \code{"top"},
\code{"top-end"}, \code{"top-start"}. Defaults to \code{bottom-start} when omitted.}

\item{size}{The dropdown's size. Must be one of \code{"large"}, \code{"medium"},
\code{"small"}. Defaults to \code{medium} when omitted.}

\item{skidding}{The offset of the dropdown menu along its trigger.
Defaults to \code{0} when omitted.}

\item{trigger}{The element that triggers the dropdown, such as a
\verb{<wa-button>} or \verb{<button>}.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-dropdown} component.
Generated from Web Awesome metadata.
}
