% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_dropdown_item.R
\name{wa_dropdown_item}
\alias{wa_dropdown_item}
\title{Create a \code{wa-dropdown-item} component}
\usage{
wa_dropdown_item(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  value = NULL,
  checked = NULL,
  disabled = NULL,
  dir = NULL,
  lang = NULL,
  submenu_open = NULL,
  type = NULL,
  variant = NULL,
  details = NULL,
  icon = NULL,
  submenu = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{String. An optional value for the menu item. This is
useful for determining which item was selected when listening to the
dropdown's \code{wa-select} event.}

\item{checked}{Boolean. Default: \code{FALSE}. Set to TRUE to check the
dropdown item. Only valid when \code{type} is \code{checkbox}.}

\item{disabled}{Boolean. Default: \code{FALSE}. Disables the dropdown item.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{submenu_open}{Boolean. Default: \code{FALSE}. Whether the submenu is
currently open.}

\item{type}{Enumerated string. Allowed values: \code{checkbox}, \code{normal}.
Default: \code{normal}. Set to \code{checkbox} to make the item a checkbox.}

\item{variant}{Enumerated string. Allowed values: \code{danger}, \code{default}.
Default: \code{default}. The type of menu item to render.}

\item{details}{Additional content or details to display after the
label.}

\item{icon}{An optional icon to display before the label.}

\item{submenu}{Submenu items, typically \verb{<wa-dropdown-item>} elements,
to create a nested menu.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-dropdown-item} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

