% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_format_bytes.R
\name{wa_format_bytes}
\alias{wa_format_bytes}
\title{Create a \code{wa-format-bytes} component}
\usage{
wa_format_bytes(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  value = NULL,
  dir = NULL,
  display = NULL,
  lang = NULL,
  unit = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{Number. Default: \code{0}. The number to format in bytes.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{display}{Enumerated string. Allowed values: \code{long}, \code{narrow},
\code{short}. Default: \code{short}. Determines how to display the result, e.g.
"100 bytes", "100 b", or "100b".}

\item{lang}{String. Optional Web Awesome attribute.}

\item{unit}{Enumerated string. Allowed values: \code{bit}, \code{byte}. Default:
\code{byte}. The type of unit to display.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-format-bytes} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

