% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_format_date.R
\name{wa_format_date}
\alias{wa_format_date}
\title{Create a \code{wa-format-date} component}
\usage{
wa_format_date(
  ...,
  id = NULL,
  date = NULL,
  day = NULL,
  dir = NULL,
  era = NULL,
  hour = NULL,
  hour_format = NULL,
  lang = NULL,
  minute = NULL,
  month = NULL,
  second = NULL,
  time_zone = NULL,
  time_zone_name = NULL,
  weekday = NULL,
  year = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{date}{The date/time to format. If not set, the current date and
time will be used. When passing a string, it's strongly recommended to
use the ISO 8601 format to ensure timezones are handled correctly. To
convert a date to this format in JavaScript, use \code{date.toISOString()}.
Defaults to \verb{new Date()} when omitted.}

\item{day}{The format for displaying the day. Must be one of
\code{"2-digit"}, \code{"numeric"}.}

\item{dir}{Optional Web Awesome attribute.}

\item{era}{The format for displaying the era. Must be one of \code{"long"},
\code{"narrow"}, \code{"short"}.}

\item{hour}{The format for displaying the hour. Must be one of
\code{"2-digit"}, \code{"numeric"}.}

\item{hour_format}{The format for displaying the hour. Must be one of
\code{"12"}, \code{"24"}, \code{"auto"}. Defaults to \code{auto} when omitted.}

\item{lang}{Optional Web Awesome attribute.}

\item{minute}{The format for displaying the minute. Must be one of
\code{"2-digit"}, \code{"numeric"}.}

\item{month}{The format for displaying the month. Must be one of
\code{"2-digit"}, \code{"long"}, \code{"narrow"}, \code{"numeric"}, \code{"short"}.}

\item{second}{The format for displaying the second. Must be one of
\code{"2-digit"}, \code{"numeric"}.}

\item{time_zone}{The time zone to express the time in.}

\item{time_zone_name}{The format for displaying the time. Must be one
of \code{"long"}, \code{"short"}.}

\item{weekday}{The format for displaying the weekday. Must be one of
\code{"long"}, \code{"narrow"}, \code{"short"}.}

\item{year}{The format for displaying the year. Must be one of
\code{"2-digit"}, \code{"numeric"}.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-format-date} component.
Generated from Web Awesome metadata.
}
