% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_format_number.R
\name{wa_format_number}
\alias{wa_format_number}
\title{Create a \code{wa-format-number} component}
\usage{
wa_format_number(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  value = NULL,
  currency = NULL,
  currency_display = NULL,
  dir = NULL,
  lang = NULL,
  maximum_fraction_digits = NULL,
  maximum_significant_digits = NULL,
  minimum_fraction_digits = NULL,
  minimum_integer_digits = NULL,
  minimum_significant_digits = NULL,
  type = NULL,
  without_grouping = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{Number. Default: \code{0}. The number to format.}

\item{currency}{String. Default: \code{USD}. The ISO 4217 currency code to
use when formatting.}

\item{currency_display}{Enumerated string. Allowed values: \code{code},
\code{name}, \code{narrowSymbol}, \code{symbol}. Default: \code{symbol}. How to display the
currency.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{maximum_fraction_digits}{Number. The maximum number of fraction
digits to use. Possible values are 0-100.}

\item{maximum_significant_digits}{Number. The maximum number of
significant digits to use,. Possible values are 1-21.}

\item{minimum_fraction_digits}{Number. The minimum number of fraction
digits to use. Possible values are 0-100.}

\item{minimum_integer_digits}{Number. The minimum number of integer
digits to use. Possible values are 1-21.}

\item{minimum_significant_digits}{Number. The minimum number of
significant digits to use. Possible values are 1-21.}

\item{type}{Enumerated string. Allowed values: \code{currency}, \code{decimal},
\code{percent}. Default: \code{decimal}. The formatting style to use.}

\item{without_grouping}{Boolean. Default: \code{FALSE}. Turns off grouping
separators.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-format-number} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

