% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_include.R
\name{wa_include}
\alias{wa_include}
\title{Create a \code{wa-include} component}
\usage{
wa_include(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  allow_scripts = NULL,
  dir = NULL,
  lang = NULL,
  mode = NULL,
  src = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{allow_scripts}{Allows included scripts to be executed. Be sure
you trust the content you are including as it will be executed as code
and can result in XSS attacks. Defaults to \code{false} when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{mode}{The fetch mode to use. Must be one of \code{"cors"},
\code{"no-cors"}, \code{"same-origin"}. Defaults to \code{cors} when omitted.}

\item{src}{The location of the HTML file to include. Be sure you trust
the content you are including as it will be executed as code and can
result in XSS attacks.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-include} component.
Generated from Web Awesome metadata.
}
