% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_mutation_observer.R
\name{wa_mutation_observer}
\alias{wa_mutation_observer}
\title{Create a \code{wa-mutation-observer} component}
\usage{
wa_mutation_observer(
  ...,
  id = NULL,
  disabled = NULL,
  attr = NULL,
  attr_old_value = NULL,
  char_data = NULL,
  char_data_old_value = NULL,
  child_list = NULL,
  dir = NULL,
  lang = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{disabled}{Disables the observer. Defaults to \code{false} when
omitted.}

\item{attr}{Watches for changes to attributes. To watch only specific
attributes, separate them by a space, e.g. \code{attr="class id title"}. To
watch all attributes, use \code{*}.}

\item{attr_old_value}{Indicates whether or not the attribute's previous
value should be recorded when monitoring changes. Defaults to \code{false}
when omitted.}

\item{char_data}{Watches for changes to the character data contained
within the node. Defaults to \code{false} when omitted.}

\item{char_data_old_value}{Indicates whether or not the previous value
of the node's text should be recorded. Defaults to \code{false} when
omitted.}

\item{child_list}{Watches for the addition or removal of new child
nodes. Defaults to \code{false} when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-mutation-observer} component.
Generated from Web Awesome metadata.
}
