% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_number_input.R
\name{wa_number_input}
\alias{wa_number_input}
\alias{update_wa_number_input}
\title{Create a \code{wa-number-input} component}
\usage{
wa_number_input(
  input_id,
  ...,
  value = NULL,
  disabled = NULL,
  label = NULL,
  hint = NULL,
  name = NULL,
  appearance = NULL,
  autocomplete = NULL,
  autofocus = NULL,
  custom_error = NULL,
  dir = NULL,
  enterkeyhint = NULL,
  inputmode = NULL,
  lang = NULL,
  max = NULL,
  min = NULL,
  pill = NULL,
  placeholder = NULL,
  readonly = NULL,
  required = NULL,
  size = NULL,
  step = NULL,
  title = NULL,
  with_hint = NULL,
  with_label = NULL,
  without_steppers = NULL,
  decrement_icon = NULL,
  end = NULL,
  hint_slot = NULL,
  increment_icon = NULL,
  label_slot = NULL,
  start = NULL
)

update_wa_number_input(
  session,
  input_id,
  value = NULL,
  label = NULL,
  hint = NULL,
  disabled = NULL
)
}
\arguments{
\item{input_id}{Input id of the component to update.}

\item{...}{Child content for the component's default slot.}

\item{value}{Optional value to send to the component.}

\item{disabled}{Optional logical disabled state to send to the component.}

\item{label}{Optional label text to send to the component.}

\item{hint}{Optional hint text to send to the component.}

\item{name}{The name of the input, submitted as a name/value pair with
form data. Defaults to \code{null} when omitted.}

\item{appearance}{The input's visual appearance. Must be one of
\code{"filled"}, \code{"filled-outlined"}, \code{"outlined"}. Defaults to \code{outlined}
when omitted.}

\item{autocomplete}{Specifies what permission the browser has to
provide assistance in filling out form field values. Refer to this page
on MDN for available values.}

\item{autofocus}{Indicates that the input should receive focus on page
load.}

\item{custom_error}{Optional Web Awesome attribute. Defaults to \code{null}
when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{enterkeyhint}{Used to customize the label or icon of the Enter
key on virtual keyboards. Must be one of \code{"done"}, \code{"enter"}, \code{"go"},
\code{"next"}, \code{"previous"}, \code{"search"}, \code{"send"}.}

\item{inputmode}{Tells the browser what type of data will be entered by
the user, allowing it to display the appropriate virtual keyboard on
supportive devices. Must be one of \code{"decimal"}, \code{"numeric"}. Defaults
to \code{numeric} when omitted.}

\item{lang}{Optional Web Awesome attribute.}

\item{max}{The input's maximum value.}

\item{min}{The input's minimum value.}

\item{pill}{Draws a pill-style input with rounded edges. Defaults to
\code{false} when omitted.}

\item{placeholder}{Placeholder text to show as a hint when the input is
empty. Defaults to `` when omitted.}

\item{readonly}{Makes the input readonly. Defaults to \code{false} when
omitted.}

\item{required}{Makes the input a required field. Defaults to \code{false}
when omitted.}

\item{size}{The input's size. Must be one of \code{"large"}, \code{"medium"},
\code{"small"}. Defaults to \code{medium} when omitted.}

\item{step}{Specifies the granularity that the value must adhere to, or
the special value \code{any} which means no stepping is implied, allowing
any numeric value. Defaults to \code{1} when omitted.}

\item{title}{Optional Web Awesome attribute. Defaults to `` when
omitted.}

\item{with_hint}{Used for SSR. Will determine if the SSRed component
will have the hint slot rendered on initial paint. Defaults to \code{false}
when omitted.}

\item{with_label}{Used for SSR. Will determine if the SSRed component
will have the label slot rendered on initial paint. Defaults to \code{false}
when omitted.}

\item{without_steppers}{Hides the increment/decrement stepper buttons.
Defaults to \code{false} when omitted.}

\item{decrement_icon}{An icon to use in lieu of the default decrement
icon.}

\item{end}{An element, such as \verb{<wa-icon>}, placed at the end of the
input control (before steppers).}

\item{hint_slot}{Text that describes how to use the input.
Alternatively, you can use the \code{hint} attribute.}

\item{increment_icon}{An icon to use in lieu of the default increment
icon.}

\item{label_slot}{The input's label. Alternatively, you can use the
\code{label} attribute.}

\item{start}{An element, such as \verb{<wa-icon>}, placed at the start of
the input control.}

\item{session}{Shiny session object.}
}
\value{
An HTML tag for the component.

Invisibly returns \code{NULL}.
}
\description{
Generated wrapper for the Web Awesome \code{wa-number-input} component.
Generated from Web Awesome metadata.

Generated update helper for the Web Awesome \code{wa-number-input} component.
}
