% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_popover.R
\name{wa_popover}
\alias{wa_popover}
\title{Create a \code{wa-popover} component}
\usage{
wa_popover(
  ...,
  id = NULL,
  dir = NULL,
  distance = NULL,
  `for` = NULL,
  lang = NULL,
  open = NULL,
  placement = NULL,
  skidding = NULL,
  without_arrow = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{dir}{Optional Web Awesome attribute.}

\item{distance}{The distance in pixels from which to offset the popover
away from its target.}

\item{for}{The ID of the popover's anchor element. This must be an
interactive/focusable element such as a button.}

\item{lang}{Optional Web Awesome attribute.}

\item{open}{Shows or hides the popover.}

\item{placement}{The preferred placement of the popover. Note that the
actual placement may vary as needed to keep the popover inside of the
viewport.}

\item{skidding}{The distance in pixels from which to offset the popover
along its target.}

\item{without_arrow}{Removes the arrow from the popover.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-popover} component.
Generated from Web Awesome metadata.
}
