% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_popover.R
\name{wa_popover}
\alias{wa_popover}
\title{Create a \code{wa-popover} component}
\usage{
wa_popover(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  dir = NULL,
  distance = NULL,
  `for` = NULL,
  lang = NULL,
  open = NULL,
  placement = NULL,
  skidding = NULL,
  without_arrow = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{distance}{Number. Default: \code{8}. The distance in pixels from which
to offset the popover away from its target.}

\item{for}{String. Default: \code{null}. The ID of the popover's anchor
element. This must be an interactive/focusable element such as a
button.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{open}{Boolean. Default: \code{FALSE}. Shows or hides the popover.}

\item{placement}{Enumerated string. Allowed values: \code{bottom},
\code{bottom-end}, \code{bottom-start}, \code{left}, \code{left-end}, \code{left-start},
\code{right}, \code{right-end}, \code{right-start}, \code{top}, \code{top-end}, \code{top-start}.
Default: \code{top}. The preferred placement of the popover. Note that the
actual placement may vary as needed to keep the popover inside of the
viewport.}

\item{skidding}{Number. Default: \code{0}. The distance in pixels from which
to offset the popover along its target.}

\item{without_arrow}{Boolean. Default: \code{FALSE}. Removes the arrow from
the popover.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-popover} component.
Generated from Web Awesome metadata.
}
