% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_progress_bar.R
\name{wa_progress_bar}
\alias{wa_progress_bar}
\title{Create a \code{wa-progress-bar} component}
\usage{
wa_progress_bar(
  ...,
  id = NULL,
  value = NULL,
  label = NULL,
  dir = NULL,
  indeterminate = NULL,
  lang = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{value}{The current progress as a percentage, 0 to 100. Defaults
to \code{0} when omitted.}

\item{label}{A custom label for assistive devices. Defaults to `` when
omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{indeterminate}{When true, percentage is ignored, the label is
hidden, and the progress bar is drawn in an indeterminate state.
Defaults to \code{false} when omitted.}

\item{lang}{Optional Web Awesome attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-progress-bar} component.
Generated from Web Awesome metadata.
}
