% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_qr_code.R
\name{wa_qr_code}
\alias{wa_qr_code}
\title{Create a \code{wa-qr-code} component}
\usage{
wa_qr_code(
  ...,
  id = NULL,
  value = NULL,
  label = NULL,
  background = NULL,
  dir = NULL,
  error_correction = NULL,
  fill = NULL,
  lang = NULL,
  radius = NULL,
  size = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{value}{The QR code's value. Defaults to `` when omitted.}

\item{label}{The label for assistive devices to announce. If
unspecified, the value will be used instead. Defaults to `` when
omitted.}

\item{background}{The background color. This can be any valid CSS color
or \code{transparent}. It cannot be a CSS custom property. Defaults to ``
when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{error_correction}{The level of error correction to use. Learn
more Must be one of \code{"H"}, \code{"L"}, \code{"M"}, \code{"Q"}. Defaults to \code{H} when
omitted.}

\item{fill}{The fill color. This can be any valid CSS color, but not a
CSS custom property. Defaults to `` when omitted.}

\item{lang}{Optional Web Awesome attribute.}

\item{radius}{The edge radius of each module. Must be between 0 and
0.5. Defaults to \code{0} when omitted.}

\item{size}{The size of the QR code, in pixels. Defaults to \code{128} when
omitted.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-qr-code} component.
Generated from Web Awesome metadata.
}
