% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_radio.R
\name{wa_radio}
\alias{wa_radio}
\title{Create a \code{wa-radio} component}
\usage{
wa_radio(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  value = NULL,
  disabled = NULL,
  name = NULL,
  appearance = NULL,
  custom_error = NULL,
  dir = NULL,
  lang = NULL,
  size = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{String. The radio's value. When selected, the radio group
will receive this value.}

\item{disabled}{Boolean. Default: \code{FALSE}. Disables the radio.}

\item{name}{String. Default: \code{null}. The name of the input, submitted
as a name/value pair with form data.}

\item{appearance}{Enumerated string. Allowed values: \code{button},
\code{default}. Default: \code{default}. The radio's visual appearance.}

\item{custom_error}{String. Default: \code{null}. Optional Web Awesome
attribute.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{size}{Enumerated string. Allowed values: \code{large}, \code{medium},
\code{small}. The radio's size. When used inside a radio group, the size
will be determined by the radio group's size, which will override this
attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-radio} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

