% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_radio_group.R
\name{wa_radio_group}
\alias{wa_radio_group}
\title{Create a \code{wa-radio-group} component}
\usage{
wa_radio_group(
  input_id,
  ...,
  class = NULL,
  style = NULL,
  value = NULL,
  disabled = NULL,
  label = NULL,
  hint = NULL,
  name = NULL,
  custom_error = NULL,
  dir = NULL,
  lang = NULL,
  orientation = NULL,
  required = NULL,
  size = NULL,
  with_hint = NULL,
  with_label = NULL,
  hint_slot = NULL,
  label_slot = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{The default value of the form control. Primarily used for
resetting the form control. This wrapper argument sets the HTML \code{value}
attribute, which maps to the component's \code{defaultValue} field/property
rather than its live \code{value} property.}

\item{disabled}{Disables the radio group and all child radios. Defaults
to \code{false} when omitted.}

\item{label}{The radio group's label. Required for proper
accessibility. If you need to display HTML, use the \code{label} slot
instead. Defaults to `` when omitted.}

\item{hint}{The radio groups's hint. If you need to display HTML, use
the \code{hint} slot instead. Defaults to `` when omitted.}

\item{name}{The name of the radio group, submitted as a name/value pair
with form data. Defaults to \code{null} when omitted.}

\item{custom_error}{Optional Web Awesome attribute. Defaults to \code{null}
when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{orientation}{The orientation in which to show radio items. Must
be one of \code{"horizontal"}, \code{"vertical"}. Defaults to \code{vertical} when
omitted.}

\item{required}{Ensures a child radio is checked before allowing the
containing form to submit. Defaults to \code{false} when omitted.}

\item{size}{The radio group's size. When present, this size will be
applied to all \verb{<wa-radio>} items inside. Must be one of \code{"large"},
\code{"medium"}, \code{"small"}.}

\item{with_hint}{Used for SSR. if true, will show slotted hint on
initial render. Defaults to \code{false} when omitted.}

\item{with_label}{Used for SSR. if true, will show slotted label on
initial render. Defaults to \code{false} when omitted.}

\item{hint_slot}{Text that describes how to use the radio group.
Alternatively, you can use the \code{hint} attribute.}

\item{label_slot}{The radio group's label. Required for proper
accessibility. Alternatively, you can use the \code{label} attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-radio-group} component.
Generated from Web Awesome metadata.
}
