% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_rating.R
\name{wa_rating}
\alias{wa_rating}
\title{Create a \code{wa-rating} component}
\usage{
wa_rating(
  input_id,
  ...,
  value = NULL,
  disabled = NULL,
  label = NULL,
  dir = NULL,
  get_symbol = NULL,
  lang = NULL,
  max = NULL,
  precision = NULL,
  readonly = NULL,
  size = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{value}{The current rating.}

\item{disabled}{Disables the rating.}

\item{label}{A label that describes the rating to assistive devices.}

\item{dir}{Optional Web Awesome attribute.}

\item{get_symbol}{A function that customizes the symbol to be rendered.
The first and only argument is the rating's current value. The function
should return a string containing trusted HTML of the symbol to render
at the specified value. Works well with \verb{<wa-icon>} elements.}

\item{lang}{Optional Web Awesome attribute.}

\item{max}{The highest rating to show.}

\item{precision}{The precision at which the rating will increase and
decrease. For example, to allow half-star ratings, set this attribute
to \code{0.5}.}

\item{readonly}{Makes the rating readonly.}

\item{size}{The component's size.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-rating} component.
Generated from Web Awesome metadata.
}
