% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_scroller.R
\name{wa_scroller}
\alias{wa_scroller}
\title{Create a \code{wa-scroller} component}
\usage{
wa_scroller(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  dir = NULL,
  lang = NULL,
  orientation = NULL,
  without_scrollbar = NULL,
  without_shadow = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{orientation}{Enumerated string. Allowed values: \code{horizontal},
\code{vertical}. Default: \code{horizontal}. The scroller's orientation.}

\item{without_scrollbar}{Boolean. Default: \code{FALSE}. Removes the visible
scrollbar.}

\item{without_shadow}{Boolean. Default: \code{FALSE}. Removes the shadows.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-scroller} component.
Generated from Web Awesome metadata.
}
