% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_slider.R
\name{wa_slider}
\alias{wa_slider}
\alias{update_wa_slider}
\title{Create a \code{wa-slider} component}
\usage{
wa_slider(
  input_id,
  ...,
  value = NULL,
  disabled = NULL,
  label = NULL,
  hint = NULL,
  name = NULL,
  autofocus = NULL,
  custom_error = NULL,
  dir = NULL,
  indicator_offset = NULL,
  lang = NULL,
  max = NULL,
  max_value = NULL,
  min = NULL,
  min_value = NULL,
  orientation = NULL,
  range = NULL,
  readonly = NULL,
  required = NULL,
  size = NULL,
  step = NULL,
  tooltip_distance = NULL,
  tooltip_placement = NULL,
  with_markers = NULL,
  with_tooltip = NULL,
  hint_slot = NULL,
  label_slot = NULL,
  reference = NULL
)

update_wa_slider(
  session,
  input_id,
  value = NULL,
  label = NULL,
  hint = NULL,
  disabled = NULL
)
}
\arguments{
\item{input_id}{Input id of the component to update.}

\item{...}{Child content for the component's default slot.}

\item{value}{Optional value to send to the component.}

\item{disabled}{Optional logical disabled state to send to the component.}

\item{label}{Optional label text to send to the component.}

\item{hint}{Optional hint text to send to the component.}

\item{name}{The name of the slider. This will be submitted with the
form as a name/value pair.}

\item{autofocus}{Tells the browser to focus the slider when the page
loads or a dialog is shown.}

\item{custom_error}{Optional Web Awesome attribute.}

\item{dir}{Optional Web Awesome attribute.}

\item{indicator_offset}{The starting value from which to draw the
slider's fill, which is based on its current value.}

\item{lang}{Optional Web Awesome attribute.}

\item{max}{The maximum value allowed.}

\item{max_value}{The maximum value of a range selection. Used only when
range attribute is set.}

\item{min}{The minimum value allowed.}

\item{min_value}{The minimum value of a range selection. Used only when
range attribute is set.}

\item{orientation}{The orientation of the slider.}

\item{range}{Converts the slider to a range slider with two thumbs.}

\item{readonly}{Makes the slider a read-only field.}

\item{required}{Makes the slider a required field.}

\item{size}{The slider's size.}

\item{step}{The granularity the value must adhere to when incrementing
and decrementing.}

\item{tooltip_distance}{The distance of the tooltip from the slider's
thumb.}

\item{tooltip_placement}{The placement of the tooltip in reference to
the slider's thumb.}

\item{with_markers}{Draws markers at each step along the slider.}

\item{with_tooltip}{Draws a tooltip above the thumb when the control
has focus or is dragged.}

\item{hint_slot}{Text that describes how to use the input.
Alternatively, you can use the \code{hint} attribute. instead.}

\item{label_slot}{The slider label. Alternatively, you can use the
\code{label} attribute.}

\item{reference}{One or more reference labels to show visually below
the slider.}

\item{session}{Shiny session object.}
}
\value{
An HTML tag for the component.

Invisibly returns \code{NULL}.
}
\description{
\if{html}{\out{
<wa-slider>
Generated from Web Awesome metadata.
}}

Generated update helper for the Web Awesome \code{wa-slider} component.
}
