% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_switch.R
\name{wa_switch}
\alias{wa_switch}
\title{Create a \code{wa-switch} component}
\usage{
wa_switch(
  input_id,
  ...,
  class = NULL,
  style = NULL,
  value = NULL,
  checked = NULL,
  disabled = NULL,
  hint = NULL,
  name = NULL,
  custom_error = NULL,
  dir = NULL,
  lang = NULL,
  required = NULL,
  size = NULL,
  title = NULL,
  with_hint = NULL,
  hint_slot = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{value}{String. The value of the switch, submitted as a name/value
pair with form data.}

\item{checked}{Boolean. Default: \code{FALSE}. The default value of the form
control. Primarily used for resetting the form control. This wrapper
argument sets the HTML \code{checked} attribute, which maps to the
component's \code{defaultChecked} field/property rather than its live
\code{checked} property.}

\item{disabled}{Boolean. Default: \code{FALSE}. Disables the switch.}

\item{hint}{String. Default: \code{""}. The switch's hint. If you need to
display HTML, use the \code{hint} slot instead.}

\item{name}{String. Default: \code{null}. The name of the switch, submitted
as a name/value pair with form data.}

\item{custom_error}{String. Default: \code{null}. Optional Web Awesome
attribute.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{required}{Boolean. Default: \code{FALSE}. Makes the switch a required
field.}

\item{size}{Enumerated string. Allowed values: \code{large}, \code{medium},
\code{small}. Default: \code{medium}. The switch's size.}

\item{title}{String. Default: \code{""}. Optional Web Awesome attribute.}

\item{with_hint}{Boolean. Default: \code{FALSE}. Used for SSR. If you slot
in hint, make sure to add \code{with-hint} to your component to get it to
properly render with SSR.}

\item{hint_slot}{Text that describes how to use the switch.
Alternatively, you can use the \code{hint} attribute.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-switch} component. When used
as a Shiny input, \verb{input$<input_id>} reflects the component's current
\code{value} value. The Shiny value is returned as a logical value.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

\verb{input$<input_id>} reflects the component's current \code{value} value. The
Shiny value is returned as a logical value.
}

