% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_tab.R
\name{wa_tab}
\alias{wa_tab}
\title{Create a \code{wa-tab} component}
\usage{
wa_tab(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  disabled = NULL,
  dir = NULL,
  lang = NULL,
  panel = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{disabled}{Disables the tab and prevents selection. Defaults to
\code{false} when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{panel}{The name of the tab panel this tab is associated with. The
panel must be located in the same tab group. Defaults to `` when
omitted.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-tab} component.
Generated from Web Awesome metadata.
}
