% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_tab_group.R
\name{wa_tab_group}
\alias{wa_tab_group}
\title{Create a \code{wa-tab-group} component}
\usage{
wa_tab_group(
  input_id,
  ...,
  class = NULL,
  style = NULL,
  activation = NULL,
  active = NULL,
  dir = NULL,
  lang = NULL,
  placement = NULL,
  without_scroll_controls = NULL,
  nav = NULL
)
}
\arguments{
\item{input_id}{Shiny input id for the component. This is also used as
the rendered DOM \code{id} attribute.}

\item{...}{Child content for the component's default slot.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{activation}{When set to auto, navigating tabs with the arrow keys
will instantly show the corresponding tab panel. When set to manual,
the tab will receive focus but will not show until the user presses
spacebar or enter. Must be one of \code{"auto"}, \code{"manual"}. Defaults to
\code{auto} when omitted.}

\item{active}{Sets the active tab. Defaults to `` when omitted.}

\item{dir}{Optional Web Awesome attribute.}

\item{lang}{Optional Web Awesome attribute.}

\item{placement}{The placement of the tabs. Must be one of \code{"bottom"},
\code{"end"}, \code{"start"}, \code{"top"}. Defaults to \code{top} when omitted.}

\item{without_scroll_controls}{Disables the scroll arrows that appear
when tabs overflow. Defaults to \code{false} when omitted.}

\item{nav}{Used for grouping tabs in the tab group. Must be \verb{<wa-tab>}
elements. Note that \verb{<wa-tab>} will set this slot on itself
automatically.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-tab-group} component.
Generated from Web Awesome metadata.
}
