% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_tag.R
\name{wa_tag}
\alias{wa_tag}
\title{Create a \code{wa-tag} component}
\usage{
wa_tag(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  appearance = NULL,
  dir = NULL,
  lang = NULL,
  pill = NULL,
  size = NULL,
  variant = NULL,
  with_remove = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{appearance}{Enumerated string. Allowed values: \code{accent},
\code{filled}, \code{filled-outlined}, \code{outlined}. Default: \code{filled-outlined}.
The tag's visual appearance.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{pill}{Boolean. Default: \code{FALSE}. Draws a pill-style tag with
rounded edges.}

\item{size}{Enumerated string. Allowed values: \code{large}, \code{medium},
\code{small}. Default: \code{medium}. The tag's size.}

\item{variant}{Enumerated string. Allowed values: \code{brand}, \code{danger},
\code{neutral}, \code{success}, \code{warning}. Default: \code{neutral}. The tag's theme
variant. Defaults to \code{neutral} if not within another element with a
variant.}

\item{with_remove}{Boolean. Default: \code{FALSE}. Makes the tag removable
and shows a remove button.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-tag} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

