% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_tooltip.R
\name{wa_tooltip}
\alias{wa_tooltip}
\title{Create a \code{wa-tooltip} component}
\usage{
wa_tooltip(
  ...,
  id = NULL,
  disabled = NULL,
  dir = NULL,
  distance = NULL,
  `for` = NULL,
  hide_delay = NULL,
  lang = NULL,
  open = NULL,
  placement = NULL,
  show_delay = NULL,
  skidding = NULL,
  trigger = NULL,
  without_arrow = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{disabled}{Disables the tooltip so it won't show when triggered.}

\item{dir}{Optional Web Awesome attribute.}

\item{distance}{The distance in pixels from which to offset the tooltip
away from its target.}

\item{for}{Optional Web Awesome attribute.}

\item{hide_delay}{The amount of time to wait before hiding the tooltip
when the user mouses out.}

\item{lang}{Optional Web Awesome attribute.}

\item{open}{Indicates whether or not the tooltip is open. You can use
this in lieu of the show/hide methods.}

\item{placement}{The preferred placement of the tooltip. Note that the
actual placement may vary as needed to keep the tooltip inside of the
viewport.}

\item{show_delay}{The amount of time to wait before showing the tooltip
when the user mouses in.}

\item{skidding}{The distance in pixels from which to offset the tooltip
along its target.}

\item{trigger}{Controls how the tooltip is activated. Possible options
include \code{click}, \code{hover}, \code{focus}, and \code{manual}. Multiple options can
be passed by separating them with a space. When manual is used, the
tooltip must be activated programmatically.}

\item{without_arrow}{Removes the arrow from the tooltip.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-tooltip} component.
Generated from Web Awesome metadata.
}
