% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_tree_item.R
\name{wa_tree_item}
\alias{wa_tree_item}
\title{Create a \code{wa-tree-item} component}
\usage{
wa_tree_item(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  disabled = NULL,
  dir = NULL,
  expanded = NULL,
  lang = NULL,
  lazy = NULL,
  selected = NULL,
  collapse_icon = NULL,
  expand_icon = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{disabled}{Boolean. Default: \code{FALSE}. Disables the tree item.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{expanded}{Boolean. Default: \code{FALSE}. Expands the tree item.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{lazy}{Boolean. Default: \code{FALSE}. Enables lazy loading behavior.}

\item{selected}{Boolean. Default: \code{FALSE}. Draws the tree item in a
selected state.}

\item{collapse_icon}{The icon to show when the tree item is collapsed.}

\item{expand_icon}{The icon to show when the tree item is expanded.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-tree-item} component.
Generated from Web Awesome metadata.
}
