% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa_zoomable_frame.R
\name{wa_zoomable_frame}
\alias{wa_zoomable_frame}
\title{Create a \code{wa-zoomable-frame} component}
\usage{
wa_zoomable_frame(
  ...,
  id = NULL,
  class = NULL,
  style = NULL,
  allowfullscreen = NULL,
  dir = NULL,
  lang = NULL,
  loading = NULL,
  referrerpolicy = NULL,
  sandbox = NULL,
  src = NULL,
  srcdoc = NULL,
  with_theme_sync = NULL,
  without_controls = NULL,
  without_interaction = NULL,
  zoom = NULL,
  zoom_levels = NULL,
  zoom_in_icon = NULL,
  zoom_out_icon = NULL
)
}
\arguments{
\item{...}{Child content for the component's default slot.}

\item{id}{Optional DOM id attribute for HTML, CSS, and JS targeting.}

\item{class}{Optional CSS class string.}

\item{style}{Optional inline CSS style string.}

\item{allowfullscreen}{Boolean. Default: \code{FALSE}. Allows fullscreen
mode.}

\item{dir}{String. Optional Web Awesome attribute.}

\item{lang}{String. Optional Web Awesome attribute.}

\item{loading}{Enumerated string. Allowed values: \code{eager}, \code{lazy}.
Default: \code{eager}. Controls iframe loading behavior.}

\item{referrerpolicy}{String. Controls referrer information.}

\item{sandbox}{String. Security restrictions for the iframe.}

\item{src}{String. The URL of the content to display.}

\item{srcdoc}{String. Inline HTML to display.}

\item{with_theme_sync}{Boolean. Default: \code{FALSE}. Enables automatic
theme syncing (light/dark mode and theme selector classes) from the
host document to the iframe.}

\item{without_controls}{Boolean. Default: \code{FALSE}. Removes the zoom
controls.}

\item{without_interaction}{Boolean. Default: \code{FALSE}. Disables
interaction when present.}

\item{zoom}{Number. Default: \code{1}. The current zoom of the frame, e.g. 0
= 0\% and 1 = 100\%.}

\item{zoom_levels}{String. Default: \verb{25\% 50\% 75\% 100\% 125\% 150\% 175\% 200\%}. The zoom levels to step through when using zoom controls. This
does not restrict programmatic changes to the zoom.}

\item{zoom_in_icon}{The slot that contains the zoom in icon.}

\item{zoom_out_icon}{The slot that contains the zoom out icon.}
}
\value{
An HTML tag for the component.
}
\description{
Generated wrapper for the Web Awesome \code{wa-zoomable-frame} component.
Generated from Web Awesome metadata.
}
\section{Shiny Bindings}{

None.
}

